### 2022-07-18 #4576 @arfon
Note to editors this is a resubmission of https://github.com/openjournals/joss-reviews/issues/3116
