### 2022-07-18 f2d5efe
Added JOSS batch in README
