### 2022-07-18 2c25af1
Minor correction of paper due to compiling issues with JOSS
