+++ PAPER_JOSS/paper.md
@@ -21,7 +21,7 @@ authors:
     orcid: 0000-0003-3341-4085
     affiliation: 1
   - name: Daniel Forstenlechner
-  	affiliation: 1
+    affiliation: 1
   - name: Georg S. Graßler
     affiliation: 1   
   - name:  Sergei V. Kasilov
