### 2022-07-19 #4576 @arfon
:wave: @pibion – would you be willing to edit this submission for JOSS? Please note that this is a resubmission of https://github.com/openjournals/joss-reviews/issues/3116 so the first task would likely be to assess whether this submission is ready to head out for review.
