### 2022-08-15 08b512f
Carried over updates of APPLETS to CORE
