+++ SRC/tetra_physics_mod.f90
@@ -14,7 +14,9 @@
       !Field properties
       double precision, dimension(3)   :: curlA  ! $\epsilon^{ijk}\difp{A_k}{x^j}$ = vector a_4^i ("curl A")
       double precision :: bmod1                  ! module of B in the first vertex
+      double precision :: Atheta1                ! theta-component of the vector potential A in the first vertex
       double precision :: Aphi1                  ! phi-component of the vector potential A in the first vertex
+      double precision :: h1_1                   ! 1st component of the unit vector h in the first vertex
       double precision :: h2_1                   ! 2nd component of the unit vector h in the first vertex
       double precision :: h3_1                   ! 3rd component of the unit vector h in the first vertex
       double precision :: Phi1                   ! electrostatic potential in the first vertex
@@ -40,6 +42,7 @@
       double precision, dimension(3)   :: gR     ! gradient of R (major radius)
       double precision, dimension(3)   :: gZ     ! gradient of Z
       double precision, dimension(3)   :: gsqg   ! gradient of square root g (metric determinant) (ONLY for grid_kind = 3)
+      double precision, dimension(3)   :: gAtheta! gradient of theta-component of the vector potential A
       double precision, dimension(3)   :: gAphi  ! gradient of phi-component of the vector potential A
       double precision, dimension(3)   :: gh1    ! gradient of the 1st component of the unit vector h
       double precision, dimension(3)   :: gh2    ! gradient of the 2nd component of the unit vector h
@@ -69,6 +72,21 @@
     end type tetrahedron_skew_coord
 !
     type(tetrahedron_skew_coord), dimension(:),   allocatable, public, protected :: tetra_skew_coord
+!
+!---------------------------------------------------------------------------------------------------------
+!
+    type hamiltonian_time_type
+        sequence
+!
+        double precision                :: h1_in_curlA               !Scalar product of $\bh$ in the first vertex with curl of $\bA$ - $(h_i)_0 \epsilon^{ijk}\difp{A_k}{x^j}$
+        double precision                :: h1_in_curlh               !Scalar product of $\bh$ in the first vertex with curl of $\bh$ - $(h_i)_0 \epsilon^{ijk}\difp{A_k}{x^j}$
+        double precision, dimension(3)  :: vec_mismatch_der          !Mismatch vector: $\difp{h_i}{x^l} \epsilon^{ijk} \difp{A_k}{x^j}
+        double precision, dimension(3)  :: vec_parcurr_der           !Parallel current derivative vector: $\difp{h_i}{x^l} \epsilon^{ijk} \difp{h_k}{x^j}
+        double precision :: dt_dtau_const_save                       ! Factor dt_dtau averaged of the four vertices
+!
+    end type hamiltonian_time_type
+!
+    type(hamiltonian_time_type), dimension(:), allocatable, public, protected :: hamiltonian_time
 !
 !---------------------------------------------------------------------------------------------------------
 !
@@ -113,11 +131,17 @@
[…]
