+++ SRC/pusher_tetra_poly.f90
@@ -105,12 +105,12 @@ module pusher_tetra_poly_mod
 !ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
 !
         subroutine pusher_tetra_poly(poly_order,ind_tetr_inout,iface,x,vpar,z_save,t_remain_in,t_pass, &
-                                               & boole_t_finished,iper_phi)
+                                               & boole_t_finished,iper_phi,t_hamiltonian,gyrophase)
 !
             use tetra_physics_mod, only: tetra_physics,particle_charge,particle_mass
             use gorilla_diag_mod,only: diag_pusher_tetry_poly
             use pusher_tetra_func_mod, only: pusher_handover2neighbour
-            use gorilla_settings_mod, only: i_precomp, boole_guess
+            use gorilla_settings_mod, only: i_precomp, boole_guess, boole_time_Hamiltonian, boole_gyrophase
 !
             implicit none
 !
@@ -123,6 +123,7 @@ module pusher_tetra_poly_mod
             double precision, intent(in)                        :: t_remain_in
             double precision, intent(out)                       :: t_pass
             logical, intent(out)                                :: boole_t_finished
+            double precision, intent(out),optional              :: t_hamiltonian,gyrophase
             logical, dimension(4)                               :: boole_faces
             integer                                             :: i,j,k
             double precision, dimension(4)                      :: z,operator_b_in_b,z_dummy
@@ -136,8 +137,31 @@ module pusher_tetra_poly_mod
 !
             !Initial computation values
             z = z_init
+!
+            !Hamiltonian time (Only Delta-value within one pushing is used for output)
+            if(boole_time_Hamiltonian) then
+                if(present(t_hamiltonian)) then
+                    t_hamiltonian = 0.d0
+                else
+                    print *, 'Error: Hamiltonian-time dynamics computation is on, but NOT used. ', &
+                             & 'Please, set boole_time_Hamiltonian to .false.'
+                    stop
+                endif
+            endif
+!
+            !Gyrophase (Only Delta-value within one pushing is used for output)
+            if(boole_gyrophase) then
+                if(present(gyrophase)) then
+                    gyrophase = 0.d0
+                else
+                    print *, 'Error: Gyrophyse computation is on, but NOT used. ', &
+                             & 'Please, set boole_gyrophase to .false.'
+                    stop
+                endif
+            endif
 ! 
 if(diag_pusher_tetry_poly) then
+    print *, 'z_init', z_init
     print *, 'iface init', iface_init
     print *, 'norm at start'
     do i = 1,4
@@ -240,7 +264,15 @@ if(diag_pusher_tetry_poly) print *, 'Error in predicted integrator: Analytic app
 !
             !Integrate trajectory analytically, if root exists
             if(boole_face_correct) then
-                call analytic_integration(poly_order,i_precomp,z,tau)
+                if(boole_time_Hamiltonian) then
+                    if(boole_gyrophase) then
+                        call analytic_integration(poly_order,i_precomp,z,tau,t_hamiltonian,gyrophase)
+                    else
+                        call analytic_integration(poly_order,i_precomp,z,tau,t_hamiltonian)
+                    endif !gyrophase
+                else
+                    call analytic_integration(poly_order,i_precomp,z,tau)
+                endif
 !
                 !Validation loop ('3-planes'-control)
                 three_planes_loop: do j=1,3                             !Just consider the planes without the "exit-plane"
@@ -284,7 +316,15 @@ if(diag_pusher_tetry_poly) print *, 'Error in predicted integrator: normal veloc
                         endif
 !                    
                         !Integrate trajectory analytically
-                        call analytic_integration(poly_order,i_precomp,z,tau)
+                        if(boole_time_Hamiltonian) then
[…]
