### 2022-08-16 bb1cfef
Clarified missleading error message
