+++ SRC/pusher_tetra_poly.f90
@@ -36,10 +36,23 @@ module pusher_tetra_poly_mod
                                0.d0, 0.d0, 1.d0, 0.d0, &
                                0.d0, 0.d0, 0.d0, 1.d0 ], [4,4])
 !
-    !$OMP THREADPRIVATE(ind_tetr,iface_init,perpinv,perpinv2,dt_dtau_const,bmod0,t_remain,x_init,  &
-    !$OMP& z_init,k1,k3,vmod0)
+    !change those for adaptive step sizes, probably allocatable
+    double precision, dimension(:), allocatable           :: tau_steps_list
+    double precision, dimension(:,:), allocatable         :: intermediate_z0_list
+    integer                                               :: number_of_integration_steps
 !
-    public :: pusher_tetra_poly,initialize_const_motion_poly, &
+    !Diagnostics for adaptive step scheme (only useable for one particle calculation)
+    double precision, dimension(:,:), allocatable         :: total_fluctuation_report, single_step_fluctuation_report, &
+                                                           & closure_fluctuation_report
+    integer                                               :: report_total_entry_index, report_single_step_entry_index, & 
+                                                            & report_entry_index
+    integer, parameter                                    :: max_n_report_entries = 1, minimum_partition_number = 9000
+    logical                                               :: boole_collect_data
+!
+    !$OMP THREADPRIVATE(ind_tetr,iface_init,perpinv,perpinv2,dt_dtau_const,bmod0,t_remain,x_init,  &
+    !$OMP& z_init,k1,k3,vmod0,tau_steps_list,intermediate_z0_list,number_of_integration_steps)
+!
+    public :: pusher_tetra_poly,initialize_const_motion_poly, initialize_intermediate_steps_arrays, &
         & Quadratic_Solver2, Cubic_Solver, Quartic_Solver,analytic_integration_without_precomp,energy_tot_func
 !   
     contains
@@ -56,6 +69,66 @@ module pusher_tetra_poly_mod
             perpinv2 = perpinv2_in
 !            
         end subroutine initialize_const_motion_poly
+!
+!ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
+!
+        subroutine initialize_intermediate_steps_arrays(status)
+!
+            use gorilla_settings_mod, only: max_n_intermediate_steps, boole_adaptive_time_steps
+            use gorilla_diag_mod, only: report_pusher_tetry_poly_adaptive
+!
+            implicit none
+!            
+            integer, intent(in) :: status
+!
+            integer             :: max_entries, k
+!
+            if (status .eq. 0) then
+                !From one face to another there are max_n_intermediate_steps 
+                !and that all possible two times as trajectory can be prolonged to a third face (turning on face) + buffer
+                if (boole_adaptive_time_steps) then
+                    max_entries = 3*(max_n_intermediate_steps)
+                    allocate(tau_steps_list(max_entries),intermediate_z0_list(4,max_entries))
+                else
+                    !if no adaptive scheme -> just two steps in total
+                    allocate(tau_steps_list(2),intermediate_z0_list(4,2))
+                endif
+if (report_pusher_tetry_poly_adaptive) then
+allocate(total_fluctuation_report(max_n_report_entries*(max_n_intermediate_steps+1),2), & 
+& single_step_fluctuation_report(max_n_report_entries*(max_n_intermediate_steps+1),3), &
+& closure_fluctuation_report(max_n_report_entries*(max_n_intermediate_steps+1),3))
+report_single_step_entry_index = 0
+report_total_entry_index = 0
+report_entry_index = 0
+endif
+            elseif (status .eq. 1) then
[…]
