### 2022-08-17 6e0cd78
Updated to cleaned up version (Jonatan/Georg)
