+++ SRC/gorilla_diag_mod.f90
@@ -1,6 +1,8 @@
 !
   module gorilla_diag_mod
     logical :: diag_pusher_tetry_poly = .false.
+    logical :: diag_pusher_tetry_poly_adaptive = .false.
+    logical :: report_pusher_tetry_poly_adaptive = .false.
     logical :: diag_pusher_tetra_rk = .false.
     logical :: transp_coef_trigger = .false.
   end module gorilla_diag_mod
