+++ SRC/pusher_tetra_poly.f90
@@ -44,9 +44,8 @@ module pusher_tetra_poly_mod
     !Diagnostics for adaptive step scheme (only useable for one particle calculation)
     double precision, dimension(:,:), allocatable         :: total_fluctuation_report, single_step_fluctuation_report, &
                                                            & closure_fluctuation_report
-    integer                                               :: report_total_entry_index, report_single_step_entry_index, & 
-                                                            & report_entry_index
-    integer, parameter                                    :: max_n_report_entries = 1, minimum_partition_number = 9000
+    integer                                               :: report_entry_index, number_reports
+    integer, parameter                                    :: max_number_reports = 1, minimum_partition_number = 9000
     logical                                               :: boole_collect_data
 !
     !$OMP THREADPRIVATE(ind_tetr,iface_init,perpinv,perpinv2,dt_dtau_const,bmod0,t_remain,x_init,  &
@@ -92,28 +91,27 @@ module pusher_tetra_poly_mod
                     allocate(tau_steps_list(2),intermediate_z0_list(4,2))
                 endif
 if (report_pusher_tetry_poly_adaptive) then
-allocate(total_fluctuation_report(max_n_report_entries*(max_n_intermediate_steps+1),2), & 
-& single_step_fluctuation_report(max_n_report_entries*(max_n_intermediate_steps+1),3), &
-& closure_fluctuation_report(max_n_report_entries*(max_n_intermediate_steps+1),3))
-report_single_step_entry_index = 0
-report_total_entry_index = 0
+allocate(total_fluctuation_report(max_number_reports*(max_n_intermediate_steps+1),2), & 
+& single_step_fluctuation_report(max_number_reports*(max_n_intermediate_steps+1),3), &
+& closure_fluctuation_report(max_number_reports*(max_n_intermediate_steps+1),3))
 report_entry_index = 0
+number_reports = 0
 endif
             elseif(allocated(tau_steps_list).AND.allocated(intermediate_z0_list)) then
 !
 if (report_pusher_tetry_poly_adaptive) then
 open(123, file='./total_fluctuation_report.dat')
-do k = 1, report_total_entry_index
+do k = 1, report_entry_index
 write(123,*) total_fluctuation_report(k,:)
 end do
 close(123)
 open(123, file='./single_step_fluctuation_report.dat')
-do k = 1, report_single_step_entry_index
+do k = 1, report_entry_index
 write(123,*) single_step_fluctuation_report(k,:)
 end do
 close(123)
 open(123, file='./closure_fluctuation_report.dat')
-do k = 1, report_single_step_entry_index
+do k = 1, report_entry_index
 write(123,*) closure_fluctuation_report(k,:)
 end do
 close(123)
@@ -772,7 +770,7 @@ if(diag_pusher_tetry_poly) print*, 'Adaptive Stepsize equidistant was called'
         &  delta_energy_current, iface_out_adaptive, tau, z, boole_face_correct)
 !
         use gorilla_settings_mod, only: i_precomp, desired_delta_energy, max_n_intermediate_steps
-        use gorilla_diag_mod,only: diag_pusher_tetry_poly, diag_pusher_tetry_poly_adaptive, report_pusher_tetry_poly_adaptive
+        use gorilla_diag_mod,only: diag_pusher_tetry_poly_adaptive, report_pusher_tetry_poly_adaptive
 !
         implicit none
 !
@@ -788,88 +786,76 @@ if(diag_pusher_tetry_poly) print*, 'Adaptive Stepsize equidistant was called'
 !
         logical, dimension(4)                                 :: boole_faces
         double precision, dimension(4)                        :: z_start_adaptive
-        integer                                               :: i, k, n_intermediate_steps, limit, &
-                                                                & n_intermediate_steps_minimum
[…]
