### 2022-09-04 f2bf6b9
Tau and eta get now updated with every iteration of adaptive scheme
