### 2022-09-13 bf0baf5
Integrated allocation of intermediate step array into pusher itself
