+++ SRC/orbit_timestep_gorilla.f90
@@ -14,7 +14,7 @@ module orbit_timestep_gorilla_mod
         subroutine orbit_timestep_gorilla(x,vpar,vperp,t_step,boole_initialized,ind_tetr,iface,t_remain_out)
 !
             use pusher_tetra_rk_mod, only: find_tetra,pusher_tetra_rk,initialize_const_motion_rk
-            use pusher_tetra_poly_mod, only: pusher_tetra_poly,initialize_const_motion_poly,manage_intermediate_steps_arrays
+            use pusher_tetra_poly_mod, only: pusher_tetra_poly,initialize_const_motion_poly
             use tetra_physics_poly_precomp_mod , only: make_precomp_poly_perpinv, initialize_boole_precomp_poly_perpinv, &
                 & alloc_precomp_poly_perpinv
             use tetra_physics_mod, only: tetra_physics,particle_charge,particle_mass
@@ -71,7 +71,6 @@ module orbit_timestep_gorilla_mod
                     call initialize_const_motion_rk(perpinv,perpinv2)
                 case(2)
                     call initialize_const_motion_poly(perpinv,perpinv2)
-                    call manage_intermediate_steps_arrays()
             end select        
 !
 !             !NOT FULLY IMPLEMENTED YET: Precompute quatities dependent on perpinv
@@ -126,9 +125,6 @@ module orbit_timestep_gorilla_mod
                 endif
 !
             enddo !Loop for tetrahedron pushings
-!
-            !Deallocate intermediate_steps_arrays if pusher_poly was used
-            call manage_intermediate_steps_arrays()
 !
             !Compute vperp from position
             vperp = vperp_func(z_save,perpinv,ind_tetr_save)
