### 2022-09-19 dd88080
Major improvement of time dynamics: (1) Introduction of orbit time tracing in correct physical (Hamiltonian) time (2) Enabling negative orbit time tracing (3) Correct treatment of sqrt(g) - removed workaround of abs(sqrt(g))
