+++ SRC/pusher_tetra_rk.f90
@@ -3,6 +3,8 @@ module pusher_tetra_rk_mod
 !
     private
 !
+    integer                          :: sign_rhs
+    integer                          :: sign_t_step_save
     integer                          :: iface_init
     integer,public,protected         :: ind_tetr
     double precision                 :: B0,perpinv,perpinv2,vmod_init
@@ -22,7 +24,7 @@ module pusher_tetra_rk_mod
     double precision,parameter      :: eps_dtau_quad = 1.5d0
     double precision                 :: k1, k3
 !   
-    !$OMP THREADPRIVATE(ind_tetr,spamat,B0,perpinv,perpinv2,iface_init,dt_dtau_const,k1,k3, &
+    !$OMP THREADPRIVATE(ind_tetr,spamat,B0,perpinv,perpinv2,iface_init,dt_dtau_const,k1,k3,sign_rhs,sign_t_step_save, &
     !$OMP& dist1,dist_min,dist_max,Bvec,gradB,x_init,b,z_init,amat,anorm,t_remain,vmod_init,dtau_ref,dtau_max,dtau_quad)
 !    
     public :: pusher_tetra_rk,find_tetra,initialize_const_motion_rk,energy_tot_func
@@ -46,7 +48,7 @@ module pusher_tetra_rk_mod
 !
     subroutine initialize_pusher_tetra_rk_mod(ind_tetr_inout,x,iface,vpar,t_remain_in)
 !
-        use tetra_physics_mod, only: tetra_physics, cm_over_e,dt_dtau,coord_system
+        use tetra_physics_mod, only: tetra_physics, cm_over_e,dt_dtau,coord_system,sign_sqg
         use tetra_grid_settings_mod, only: grid_size
         use constants, only : clight,eps,pi
         use gorilla_settings_mod, only: boole_dt_dtau
@@ -61,8 +63,12 @@ module pusher_tetra_rk_mod
         logical                                    :: boole_vd_ExB,boole_vperp
 !
         t_remain = t_remain_in
-!    
+!
         ind_tetr=ind_tetr_inout           !Save the index of the tetrahedron locally
+!
+        !Sign of the right hand side of ODE - ensures that tau is ALWAYS positive inside the algorithm
+        sign_t_step_save = int(sign(1.d0,t_remain))
+        sign_rhs = sign_sqg * sign_t_step_save
 !
         z_init(1:3)=x-tetra_physics(ind_tetr)%x1       !x is the entry point of the particle into the tetrahedron in (R,phi,z)-coordinates
 !
@@ -77,6 +83,9 @@ module pusher_tetra_rk_mod
         gradB = tetra_physics(ind_tetr)%gb
         anorm = tetra_physics(ind_tetr)%anorm
         dt_dtau_const = tetra_physics(ind_tetr)%dt_dtau_const
+!
+        !Multiply with sign of rhs - ensures that tau is ALWAYS positive inside the algorithm
+        dt_dtau_const = dt_dtau_const*dble(sign_rhs)
 !    
         !Module of B at the entry point of the particle
         bmod=bmod_func(z_init(1:3))
@@ -108,6 +117,12 @@ module pusher_tetra_rk_mod
 !                
         spamat=perpinv*cm_over_e*tetra_physics(ind_tetr)%spalpmat &
             - clight* tetra_physics(ind_tetr)%spbetmat    !tr(a-Matrix)
+!
+        !Multiply A-matrix (4x4) and b with appropriate sign (which ensures that tau remains positive inside the algorithm)
+        amat = amat * dble(sign_rhs)
+        b = b * dble(sign_rhs)
+        Bvec = Bvec * dble(sign_rhs)
[…]
