+++ SRC/pusher_tetra_poly.f90
@@ -20,10 +20,16 @@ module pusher_tetra_poly_mod
 
     implicit none
 !
+    !change those for adaptive step sizes, probably allocatable
+    double precision, dimension(:), allocatable, public, protected            :: tau_steps_list
+    double precision, dimension(:,:), allocatable, public, protected          :: intermediate_z0_list
+    integer, public, protected                                                :: number_of_integration_steps
+!
     private
 !    
     integer                             :: iface_init
     integer, public, protected          :: ind_tetr
+    integer, public, protected          :: sign_rhs
     double precision                    :: perpinv2,vmod0
     double precision, public, protected :: perpinv,dt_dtau_const,bmod0
     double precision                    :: t_remain
@@ -36,11 +42,6 @@ module pusher_tetra_poly_mod
                                0.d0, 0.d0, 1.d0, 0.d0, &
                                0.d0, 0.d0, 0.d0, 1.d0 ], [4,4])
 !
-    !change those for adaptive step sizes, probably allocatable
-    double precision, dimension(:), allocatable           :: tau_steps_list
-    double precision, dimension(:,:), allocatable         :: intermediate_z0_list
-    integer                                               :: number_of_integration_steps
-!
     !Diagnostics for adaptive step scheme (only useable for one particle calculation)
     double precision, dimension(:,:), allocatable         :: total_fluctuation_report, single_step_fluctuation_report, &
                                                            & closure_fluctuation_report
@@ -49,10 +50,11 @@ module pusher_tetra_poly_mod
     logical                                               :: boole_collect_data
 !
     !$OMP THREADPRIVATE(ind_tetr,iface_init,perpinv,perpinv2,dt_dtau_const,bmod0,t_remain,x_init,  &
-    !$OMP& z_init,k1,k3,vmod0,tau_steps_list,intermediate_z0_list,number_of_integration_steps)
+    !$OMP& z_init,k1,k3,vmod0,tau_steps_list,intermediate_z0_list,number_of_integration_steps,sign_rhs)
 !
     public :: pusher_tetra_poly,initialize_const_motion_poly, &
-        & Quadratic_Solver2, Cubic_Solver, Quartic_Solver,analytic_integration_without_precomp,energy_tot_func
+        & Quadratic_Solver2, Cubic_Solver, Quartic_Solver,energy_tot_func, &
+        & set_integration_coef_manually
 !   
     contains
 !
@@ -129,7 +131,7 @@ endif
 !
         subroutine initialize_pusher_tetra_poly(ind_tetr_inout,x,iface,vpar,t_remain_in)
 !
-            use tetra_physics_mod, only: tetra_physics
+            use tetra_physics_mod, only: tetra_physics, sign_sqg
 !
             implicit none
 !
@@ -142,6 +144,9 @@ endif
 !    
             ind_tetr=ind_tetr_inout           !Save the index of the tetrahedron locally
 !
+            !Sign of the right hand side of ODE - ensures that tau is ALWAYS positive inside the algorithm
+            sign_rhs = sign_sqg * int(sign(1.d0,t_remain))
+!
             z_init(1:3)=x-tetra_physics(ind_tetr)%x1       !x is the entry point of the particle into the tetrahedron in (R,phi,z)-coordinates
 !
             z_init(4)=vpar                         !Transform to z_init: 1st vertex of tetrahdron is new origin
@@ -153,6 +158,9 @@ endif
 !
             !Tetrahedron constants
             dt_dtau_const = tetra_physics(ind_tetr)%dt_dtau_const
+!
[…]
