+++ SRC/orbit_timestep_gorilla.f90
@@ -34,6 +34,7 @@ module orbit_timestep_gorilla_mod
             logical                                         :: boole_t_finished
             integer                                         :: ind_tetr_save,iper,k
             double precision                                :: perpinv,perpinv2
+            integer                                         :: sign_t_step
 !
             !If orbit_timestep is called for the first time without grid position
             if(.not.boole_initialized) then
@@ -41,8 +42,11 @@ module orbit_timestep_gorilla_mod
                 !Check coordinate domain (optionally perform modulo operation)
                 call check_coordinate_domain(x)
 !
+                !Compute sign t_step
+                sign_t_step = int(sign(1.d0,t_step))
+!
                 !Find tetrahedron index and face index for position x
-                call find_tetra(x,vpar,vperp,ind_tetr,iface)
+                call find_tetra(x,vpar,vperp,ind_tetr,iface,sign_t_step)
 !               
                 !If particle doesn't lie inside any tetrahedron
                 if(ind_tetr.eq.-1) then

[…]
