+++ SRC/pusher_tetra_poly.f90
@@ -179,7 +179,7 @@ endif
             use gorilla_diag_mod,only: diag_pusher_tetry_poly
             use pusher_tetra_func_mod, only: pusher_handover2neighbour
             use gorilla_settings_mod, only: i_precomp, boole_guess, optional_quantities_type, boole_array_optional_quantities, &
-                                    &  boole_adaptive_time_steps
+                                    &  boole_adaptive_time_steps, i_time_tracing_option
 !
             implicit none
 !
@@ -204,6 +204,9 @@ endif
             logical                                               :: boole_analytical_approx,boole_face_correct
             logical                                               :: boole_trouble_shooting
             double precision, dimension(4,4)                      :: operator_b,operator_z_init
+            double precision, dimension(:), allocatable           :: t_hamiltonian_list
+            integer                                               :: i_step_root
+            double precision                                      :: t_remain_new
 !         
             call initialize_pusher_tetra_poly(ind_tetr_inout,x,iface,vpar,t_remain_in)
             !In case of first call of orbit integration -> inialize the intermediate_steps_array
@@ -335,7 +338,7 @@ if(diag_pusher_tetry_poly) print *, 'Error in predicted integrator: normal veloc
 if(diag_pusher_tetry_poly) print *, 'boole_face_correct',boole_face_correct
 !
             !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-            !!!!!!!!!!!!!!!!SECOND ATTEMPT WITHOUT GUESSES PLUS RESCALE!!!!!!!!!!!!!!!!!!
+            !!!!!!!!SECOND ATTEMPT WITHOUT GUESSES PLUS RESCALING IN 2nd ORDER!!!!!!!!!!!
             !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !
 if(diag_pusher_tetry_poly) print *, 'boole_face_correct',boole_face_correct
@@ -441,7 +444,33 @@ endif
             !Final processing
             x=z(1:3)+tetra_physics(ind_tetr)%x1
             vpar=z(4)
-            t_pass = tau*dt_dtau_const
+!
+            if(any(boole_array_optional_quantities)) then
+!
+                !Track Hamiltonian time for root finding operation, if necessary
+                if(i_time_tracing_option.eq.2) then
+                    allocate(t_hamiltonian_list(number_of_integration_steps+1))
+                    t_hamiltonian_list(1) = 0.d0
+                endif
+!
+                !loop over number_of_integration_steps
+                do i = 1,number_of_integration_steps
+                    call calc_optional_quantities(poly_order, intermediate_z0_list(:,i), tau_steps_list(i), optional_quantities)
+!
+                    !Track Hamiltonian time
+                    if(i_time_tracing_option.eq.2) t_hamiltonian_list(i+1) = optional_quantities%t_hamiltonian
+                enddo
+            endif
+!
+            !Compute passing time dependent on time tracing option
+            select case(i_time_tracing_option)
+                !Time tracing in 0th order - dt_dtau = const.
+                case(1)
+                    t_pass = tau*dt_dtau_const
+                !Hamiltonian time tracing with computation of polynomial
+                case(2)
+                    t_pass = optional_quantities%t_hamiltonian
+            end select
 !
[…]
