### 2022-09-19 6419855
Working on Hamiltonian time
