### 2022-09-19 2a6d744
Working version of adaptive scheme example
