### 2022-09-19 b414dee
Fixed bug in adaptive routine & made J_par compatible with intermediate step scheme
