### 2022-09-20 12e026b
Created example_7 RUN folder and MATLAB script
