+++ MATLAB/example_7.m
@@ -1,0 +1,469 @@
+%#######################################################################################################################
+% description:
+%-----------------------------------------------------------------------------------------------------------------------
+% * Compute collisionless guiding-center orbit with GORILLA for a trapped Deuterium particle with/without adaptive scheme.
+% * Use a field-aligned grid for a non-axisymmetric VMEC MHD equilibrium.
+% * Create a figure with the Poincar� plots (\varphi = 0) in cylindrical and symmetry flux coordinates.
+% * Compute the normalized parallel adiabatic invariant as a function of banana bounces.
+% * Compare energy fluctuation as function of dwell time and evolution of energy between the two schemes.
+%
+%#######################################################################################################################
+% used functions:
+%-----------------------------------------------------------------------------------------------------------------------
+% InputFile
+% NameList
+% read_in
+% load_copy
+%#######################################################################################################################
+% authors: Georg Graßler
+% created: 16.09.2022
+
+%For Construction of Demonstration
+%-----------------------------------------------------------------------------------------------------------------------
+close all
+
+%Predefinig options for calculation
+total_orbit_time = 1d1;
+desired_delta_energy = 1.0d-10;
+
+
+%Initialize used paths and files
+%-----------------------------------------------------------------------------------------------------------------------
+
+%Name of the current calculation to create folder
+name_test_case='example_7';
+
+%path of Matlab script
+path_script=pwd;
+
+%main path of GORILLA
+c=strsplit(path_script,'/');
+path_main=strjoin(c(1:end-1),'/');
+
+%path to run GORILLA code
+mkdir(path_main,['EXAMPLES/MATLAB_RUN/',name_test_case]);
+path_RUN=[path_main,'/EXAMPLES/MATLAB_RUN/',name_test_case];
+
+%path of input files (blueprints)
+path_inp_files=[path_main,'/INPUT'];
+
+%path of the used functions
+path_functions=[path_main,'/MATLAB/functions'];
+
+%define path for data and plots and create a new folder
+mkdir([path_script,'/data_plots'],name_test_case);
+path_data_plots=[path_script,'/data_plots/',name_test_case];
+
+%Change to RUN folder and clean it
+cd(path_RUN);
+
+%add path
+addpath(path_inp_files);
+addpath(path_functions);
+
+%initialize class for inputfiles
+gorilla = InputFile([path_inp_files,'/gorilla.inp']);
+gorilla_plot = InputFile([path_inp_files,'/gorilla_plot.inp']);
+tetra_grid = InputFile([path_inp_files,'/tetra_grid.inp']);
+
+%read default inputfiles
+%All variables get a default value from the blueprints
+gorilla.read();
+gorilla_plot.read();
+tetra_grid.read();
+
+
+%Set input variables for GORILLA
[…]
