### 2022-09-21 035beb5
Escaping of environment paramter in comparison.
Put environment parameter in quotation marks. This allows sucessfull
configuration regardless if the parameter is set or not. Otherwise it
would fail if not set, due to a syntax error.
