### 2022-09-21 ba18238
Added example_7.py and plotting for example_7
