+++ PYTHON/example_7.py
@@ -1,0 +1,440 @@
+#!/usr/bin/env python3.9
+# -*- coding: utf-8 -*-
+"""
+Created on Thu Sep 21 11:17:37 2022
+
+Example 6:
+ * Compute collisionless guiding-center orbit with GORILLA for a trapped Deuterium particle with/without adaptive scheme.
+ * Use a field-aligned grid for a non-axisymmetric VMEC MHD equilibrium.
+ * Create a figure with the Poincar� plots (\varphi = 0) in cylindrical and symmetry flux coordinates.
+ * Compute the normalized parallel adiabatic invariant as a function of banana bounces.
+ * Compare energy fluctuation as function of dwell time and evolution of energy between the two schemes.
+
+@author: Georg Graßler
+"""
+
+import f90nml
+import os
+import numpy as np
+import matplotlib as mpl
+import matplotlib.pyplot as plt
+mpl.rcdefaults()
+
+
+# For Construction of Demonstration
+# -----------------------------------------------------------------------------------------------------------------------
+
+# Predefinig options for calculation
+total_orbit_time = 1E1
+desired_delta_energy = 1E-10
+
+# Initialize used paths and files
+# -----------------------------------------------------------------------------------------------------------------------
+
+# Name of the current calculation to create folder
+name_test_case = 'example_7'
+
+# path of PYTHON script
+path_script = os.getcwd()
+
+# main path of GORILLA
+path_main= path_script + '/..'
+
+# path to RUN folder
+path_RUN = path_main + '/EXAMPLES/PYTHON_RUN/' + name_test_case
+if not os.path.exists(path_RUN):
+  os.makedirs(path_RUN)
+
+# path of input files (blueprints)
+path_inp_files = path_main + '/INPUT'
+
+# define path for plots and create a new folder
+name_data_folder ='/data_plots/' + name_test_case
+path_data_plots = path_script + name_data_folder
+if not os.path.exists(path_data_plots):
+  os.makedirs(path_data_plots)
+
+# Change to RUN folder and clean it
+os. chdir(path_RUN)
+if os.listdir('../' + name_test_case):
+  os.system('rm -r ../' + name_test_case + '/*')
+
+# load the namelists from the inputfiles with proper comma setting
+gorilla = f90nml.read(path_inp_files + '/gorilla.inp')
+gorilla.end_comma = True
+tetra_grid = f90nml.read(path_inp_files + '/tetra_grid.inp')
+tetra_grid.end_comma = True
+gorilla_plot = f90nml.read(path_inp_files + '/gorilla_plot.inp')
+gorilla_plot.end_comma = True
+
+
+# Set input variables for GORILLA
+# -----------------------------------------------------------------------------------------------------------------------
+
+# Input file gorilla.inp
+# All necessary variables for current calculation
+
+# Change in the electrostatic potential within the plasma volume in Gaussian units
+gorilla['gorillanml']['eps_Phi'] = 0
+
+# Coordinate system
+# 1 ... (R,phi,Z) cylindrical coordinate system
+# 2 ... (s,theta,phi) symmetry flux coordinate system
[…]
