### 2022-09-21 a1bca5c
Updated .yml files for example_7
