### 2022-09-23 3aea6db
Performed analysis of adaptive scheme with report routine
