+++ SRC/pusher_tetra_poly.f90
@@ -43,11 +43,8 @@ module pusher_tetra_poly_mod
                                0.d0, 0.d0, 0.d0, 1.d0 ], [4,4])
 !
     !Diagnostics for adaptive step scheme (only useable for one particle calculation)
-    double precision, dimension(:,:), allocatable         :: total_fluctuation_report, single_step_fluctuation_report, &
-                                                           & closure_fluctuation_report
-    integer                                               :: report_entry_index, number_reports
-    integer, parameter                                    :: max_number_reports = 1, minimum_partition_number = 9000
-    logical                                               :: boole_collect_data
+    double precision, dimension(:,:), allocatable         :: total_fluc_report, step_fluc_report
+    integer                                               :: report_index
 !
     !$OMP THREADPRIVATE(ind_tetr,iface_init,perpinv,perpinv2,dt_dtau_const,bmod0,t_remain,x_init,  &
     !$OMP& z_init,k1,k3,vmod0,tau_steps_list,intermediate_z0_list,number_of_integration_steps,sign_rhs)
@@ -94,35 +91,20 @@ module pusher_tetra_poly_mod
                     !if no adaptive scheme -> just two steps in total
                     allocate(tau_steps_list(2),intermediate_z0_list(4,2))
                 endif
-if (report_pusher_tetry_poly_adaptive) then
-allocate(total_fluctuation_report(max_number_reports*(max_n_intermediate_steps+1),2), & 
-& single_step_fluctuation_report(max_number_reports*(max_n_intermediate_steps+1),3), &
-& closure_fluctuation_report(max_number_reports*(max_n_intermediate_steps+1),3))
-report_entry_index = 0
-number_reports = 0
+!
+if(report_pusher_tetry_poly_adaptive) then
+open(124, file='./total_fluc_report.dat')
+open(118, file='./step_fluc_report.dat')
 endif
+
             elseif(option .eq. 1) then
+                deallocate(tau_steps_list,intermediate_z0_list)
 !
-if (report_pusher_tetry_poly_adaptive) then
-open(123, file='./total_fluctuation_report.dat')
-do k = 1, report_entry_index
-write(123,*) total_fluctuation_report(k,:)
-end do
-close(123)
-open(123, file='./single_step_fluctuation_report.dat')
-do k = 1, report_entry_index
-write(123,*) single_step_fluctuation_report(k,:)
-end do
-close(123)
-open(123, file='./closure_fluctuation_report.dat')
-do k = 1, report_entry_index
-write(123,*) closure_fluctuation_report(k,:)
-end do
-close(123)
-deallocate(total_fluctuation_report,single_step_fluctuation_report,closure_fluctuation_report)
+if(report_pusher_tetry_poly_adaptive) then
+close(124)
+close(118)
 endif
 !
-                deallocate(tau_steps_list,intermediate_z0_list)
             endif
 !            
         end subroutine manage_intermediate_steps_arrays
@@ -826,7 +808,7 @@ if(diag_pusher_tetry_poly)  print *, 'Error: Tau is out of safety boundary'
                                             &  iface_inout_adaptive, tau, z, boole_face_correct)
 !
         use gorilla_settings_mod, only: desired_delta_energy, max_n_intermediate_steps
-        use gorilla_diag_mod, only: diag_pusher_tetry_poly
+        use gorilla_diag_mod, only: diag_pusher_tetry_poly, report_pusher_tetry_poly_adaptive
 !
         implicit none
 !
@@ -839,13 +821,16 @@ if(diag_pusher_tetry_poly)  print *, 'Error: Tau is out of safety boundary'
         double precision, dimension(4), intent(inout)         :: z
         logical, intent(inout)                                :: boole_face_correct
 !
[…]
