+++ SETUP_ENERGY/example_7_developement.m
@@ -1,678 +1,0 @@
-%#######################################################################################################################
-% description:
-%-----------------------------------------------------------------------------------------------------------------------
-% * Compute collisionless guiding-center orbit with GORILLA for a trapped Deuterium particle with/without adaptive scheme.
-% * Use a field-aligned grid for a non-axisymmetric VMEC MHD equilibrium.
-% * Create a figure with the Poincar� plots (\varphi = 0) in cylindrical and symmetry flux coordinates.
-% * Compute the normalized parallel adiabatic invariant as a function of banana bounces.
-%
-%#######################################################################################################################
-% used functions:
-%-----------------------------------------------------------------------------------------------------------------------
-% InputFile
-% NameList
-% read_in
-% load_copy
-%#######################################################################################################################
-% authors: Georg Graßler
-% created: 16.09.2022
-
-%For Construction of Demonstration
-%-----------------------------------------------------------------------------------------------------------------------
-%control of script
-close all
-boole_recalculate_nonadaptive = false;
-boole_recalculate_adaptive = false;
-boole_recalculate_control_order4 = false;
-boole_recalculate_full_orbit = false;
-boole_recalculate_full_orbit_adaptive = false;
-boole_load_data = false;
-poly_order = 2;
-poly_order_adaptive = 2;
-total_orbit_time = 1d1;
-desired_delta_energy = 1.0d-10;
-max_n_intermediate_steps = 10000;
-
-boole_full_orbit = true;
-boole_e_tot = true;
-boole_poincare_vpar_0 = true;
-boole_J_par = true;
-
-boole_show_e_tot = true;
-boole_show_e_tot_fluctuation = false;
-boole_show_e_tot_fluctuation_full_orbit = true;
-boole_show_full_orbit = false;
-boole_show_J_par = false;
-boole_show_J_par_combined = true;
-boole_show_poincare_vpar_0_rphiz = false;
-boole_show_poincare_vpar_0_rphiz_combined = true;
-boole_show_poincare_vpar_0_sthetaphi_combined = true;
-
-n_skip_poincare = 5;
-n_skip_poincare_rphiz = 5;
-n_skip_J_par = 20;
-n_skip_e_tot_fluctuation = 10;
-n_skip_e_tot_fluctuation_full_orbit = 1;
-n_skip_e_tot = 1;
-
-% order2Color = [102,194,165]/256;
-% adaptiveColor = [141,160,203]/256;
-% order4Color = [252,141,98]/256;
-order2Color = [27,158,119]/256;
-% adaptiveColor = [117,112,179]/256;
-% order4Color = [217,95,2]/256;
-adaptiveColor = [217,95,2]/256;
-order4Color = [117,112,179]/256;
-
-MarkerPoincare = '.';
-MarkerSizePoincare = 15;
-CentralLineWidth = 2;
-MarkerEnergyFluctuation = '.';
-MarkerSizeEnergyFluctuation = 15;
-MarkerSize = 5;
-FontSize = 14;
-
-postprocessing_functions_path = 'postprocessing_functions';
-addpath(postprocessing_functions_path);
-
-%tests on lamor radius
-start_pitch_parameter = 0.3;
-n_skip_full_orbit = 1;
-energy_eV_start = 3d3;
-scaling_factor = 1/10;
-ispecies = 2;
-boole_guess = true;
-m_ratio = 1/(4*1800);
[…]
