+++ SRC/gorilla_settings_mod.f90
@@ -97,7 +97,7 @@
             call load_boole_array_optional_quantities
 !
             !Dependencies of input parameters
-            if( (i_time_tracing_option.eq.2) then
+            if( i_time_tracing_option.eq.2 ) then
                 if (.not.boole_time_Hamiltonian) ) then
                     print *, 'ERROR: When Hamiltonian time tracing is activated, set boole_time_Hamiltonian to TRUE in gorilla.inp.'
                     stop
