### 2022-09-27 641e586
backwards-compatible handling of WEST equilibria and MEPHIT
