+++ SRC/field_divB0.f90
@@ -22,6 +22,8 @@ end module field_c_mod
 !
 module field_eq_mod
   logical :: use_fpol = .true.                                      !<=18.12.18
+  logical :: skip_read = .false.
+  integer :: iaxieq = 0
   integer :: icall_eq=0
   integer :: nrad,nzet,icp,nwindow_r,nwindow_z
   real(kind=8), parameter                      :: pi=3.14159265358979d0
@@ -38,7 +40,7 @@ end module field_eq_mod
 !
 module field_mod
   integer          :: icall=0
-  integer          :: ipert,iequil,iaxieq=0
+  integer          :: ipert,iequil
   double precision :: ampl
 end module field_mod
 !
@@ -55,16 +57,37 @@ end module inthecore_mod
 !
 !cccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
 ! -----------------------------------------------------------------
+subroutine read_field_input
+  use input_files, only: iunit, gfile, pfile, convexfile, fluxdatapath
+  use field_c_mod, only: ntor, icftype
+  use field_eq_mod, only: nwindow_r, nwindow_z
+  use field_mod, only: ipert, iequil, ampl
+  use inthecore_mod, only: cutoff
+
+  open(iunit, file='field_divB0.inp')
+  read(iunit,*) ipert        ! 0=eq only, 1=vac, 2=vac+plas no derivatives,
+                             ! 3=plas+vac with derivatives
+  read(iunit,*) iequil       ! 0=perturbation alone, 1=with equilibrium
+  read(iunit,*) ampl         ! amplitude of perturbation, a.u.
+  read(iunit,*) ntor         ! number of toroidal harmonics
+  read(iunit,*) cutoff       ! inner cutoff in psi/psi_a units
+  read(iunit,*) icftype      ! type of coil file
+  read(iunit,*) gfile        ! equilibrium file
+  read(iunit,*) pfile        ! coil        file
+  read(iunit,*) convexfile   ! convex file for stretchcoords
+  read(iunit,*) fluxdatapath ! directory with data in flux coord.
+  read(iunit,*) nwindow_r    ! widow size for filtering of psi array over R
+  read(iunit,*) nwindow_z    ! widow size for filtering of psi array over Z
+  close(iunit)
+end subroutine read_field_input
+
 subroutine field(r,p,z,Br,Bp,Bz,dBrdR,dBrdp,dBrdZ   &
                 ,dBpdR,dBpdp,dBpdZ,dBzdR,dBzdp,dBzdZ)
 !
   use period_mod
-  use input_files
-  use field_c_mod,   only : ntor,icftype,icall_c
+  use field_c_mod, only: icall_c
   use field_mod
-  use inthecore_mod, only : incore,cutoff
-  use field_eq_mod, only : nwindow_r,nwindow_z
-  use tetra_grid_settings_mod, only: g_file_filename, convex_wall_filename, iaxieq_in !=> Michael Eder, 05 May 2022
+  use inthecore_mod, only: incore
 !
   implicit none
 !
@@ -72,33 +95,12 @@ subroutine field(r,p,z,Br,Bp,Bz,dBrdR,dBrdp,dBrdZ   &
                      ,dBpdR,dBpdp,dBpdZ,dBzdR,dBzdp,dBzdZ
   double precision :: rm,zm,Brc,Bpc,Bzc,dBrdRc,dBrdpc,dBrdZc   &
                      ,dBpdRc,dBpdpc,dBpdZc,dBzdRc,dBzdpc,dBzdZc
-  character*1024 :: dummy
 !
   if(icall .eq. 0) then
      icall = 1
-     open(iunit, file='field_divB0.inp')
[…]
