### 2022-09-27 e0af208
Working on Hamiltonian time
