### 2022-09-27 5bd8a85
Working on Hamiltonian time
