### 2022-09-27 eb13de3
Working on Hamiltonian time
