### 2022-09-27 8493a79
Working on Hamiltonian time
