### 2022-09-27 8afa9e9
Working on Hamiltonian time
