### 2022-09-27 ce9e56c
Working on Hamiltonian tim
