### 2022-09-27 8d94981
Working on Hamiltonian tim
