### 2022-09-30 #4576 @pibion
@sigvaldm
@Christovis
@ketch
@vickysharma0812
@robashaw

would any of you be interested in reviewing this submission for JOSS?  Reviewing will mean going though a checklist to make sure the code meets JOSS publication standards (for example, there's instructions on installation, a test suite, and contribution instructions).
