### 2022-09-30 #4576 @pibion
@micheder a comment on the paper - I'm still not seeing a comparison of similar software packages.  Perhaps I'm missing it?  Or, if there are no similar software packages, or fusion software packages that provide this particular set of calculations, that would be useful to state explicitly.
