### 2022-10-03 #4576 @micheder
@pibion - First of all, thank you very much for your comments and reviewing :)
Regarding similar software packages: For that purpose we have added the 3rd paragraph to the paper where we list various codes that solve the guiding-center equations. Furthermore, there we also state how our code differs from existing codes that solve these equations.
Do you have suggestions how we can improve this paragraph? Is there something essential that we should add?
Thank you very much for your help :)
