### 2022-10-03 1dd6b81
Introduced handling of negative sqrtg
