### 2022-10-03 4cf88ae
Working on negative sqrtg
