### 2022-10-04 0dad402
Working on negative time tracing
