### 2022-10-04 420c988
Working on Hamiltonian time
