+++ SRC/pusher_tetra_poly.f90
@@ -50,7 +50,7 @@ module pusher_tetra_poly_mod
     logical                                               :: boole_collect_data
 !
     !$OMP THREADPRIVATE(ind_tetr,iface_init,perpinv,perpinv2,dt_dtau_const,bmod0,t_remain,x_init,  &
-    !$OMP& z_init,k1,k3,vmod0,tau_steps_list,intermediate_z0_list,number_of_integration_steps)
+    !$OMP& z_init,k1,k3,vmod0,tau_steps_list,intermediate_z0_list,number_of_integration_steps,sign_rhs)
 !
     public :: pusher_tetra_poly,initialize_const_motion_poly, &
         & Quadratic_Solver2, Cubic_Solver, Quartic_Solver,analytic_integration_without_precomp,energy_tot_func
@@ -211,6 +211,7 @@ endif
             logical                                               :: boole_analytical_approx,boole_face_correct
             logical                                               :: boole_trouble_shooting
             double precision, dimension(4,4)                      :: operator_b,operator_z_init
+            double precision, allocatable                         :: t_hamiltonian
             double precision, dimension(:), allocatable           :: t_hamiltonian_list
             integer                                               :: i_step_root
             double precision                                      :: t_remain_new
@@ -452,23 +453,6 @@ endif
             x=z(1:3)+tetra_physics(ind_tetr)%x1
             vpar=z(4)
 !
-            if(any(boole_array_optional_quantities)) then
-!
-                !Track Hamiltonian time for root finding operation, if necessary
-                if(i_time_tracing_option.eq.2) then
-                    allocate(t_hamiltonian_list(number_of_integration_steps+1))
-                    t_hamiltonian_list(1) = 0.d0
-                endif
-!
-                !loop over number_of_integration_steps
-                do i = 1,number_of_integration_steps
-                    call calc_optional_quantities(poly_order, intermediate_z0_list(:,i), tau_steps_list(i), optional_quantities)
-!
-                    !Track Hamiltonian time
-                    if(i_time_tracing_option.eq.2) t_hamiltonian_list(i+1) = optional_quantities%t_hamiltonian
-                enddo
-            endif
-!
             !Compute passing time dependent on time tracing option
             select case(i_time_tracing_option)
                 !Time tracing in 0th order - dt_dtau = const.
@@ -476,7 +460,23 @@ endif
                     t_pass = tau*dt_dtau_const
                 !Hamiltonian time tracing with computation of polynomial
                 case(2)
-                    t_pass = optional_quantities%t_hamiltonian
+!
+                    !Track Hamiltonian time for root finding operation
+                    allocate(t_hamiltonian_list(number_of_integration_steps+1))
+                    t_hamiltonian_list(1) = 0.d0
+                    allocate(t_hamiltonian)
+                    t_hamiltonian = 0.d0
+!
+                    !loop over number_of_integration_steps
+                    do i = 1,number_of_integration_steps
+                        !Calculate Hamiltonian time
+                        call calc_t_hamiltonian(poly_order, intermediate_z0_list(:,i), tau_steps_list(i), t_hamiltonian)
+!
+                        !Track Hamiltonian time
+                        t_hamiltonian_list(i+1) = t_hamiltonian
+                    enddo
+!
+                    t_pass = t_hamiltonian
             end select
 !
 if(diag_pusher_tetry_poly) print *, 'tau total',tau
@@ -569,23 +569,12 @@ if(diag_pusher_tetry_poly) print *, 'tau until t finished',tau
                     x=z(1:3)+tetra_physics(ind_tetr)%x1
                     vpar=z(4)
[…]
