+++ SRC/gorilla_settings_mod.f90
@@ -98,11 +98,6 @@
 !
             !Dependencies of input parameters
             if( i_time_tracing_option.eq.2 ) then
-                if (.not.boole_time_Hamiltonian ) then
-                    print *, 'ERROR: When Hamiltonian time tracing is activated, set boole_time_Hamiltonian to TRUE in gorilla.inp.'
-                    stop
-                endif
-!
                 if (ipusher.ne.2) then
                     print *, 'ERROR: When Hamiltonian time tracing is activated, set ipusher to 2 in gorilla.inp.'
                     stop
