### 2022-10-04 4b52cec
Working on Hamiltonian time
