+++ SRC/pusher_tetra_poly.f90
@@ -534,6 +534,7 @@ if(diag_pusher_tetry_poly) print *, 't_pass',t_pass
 !
                         !Fail save: consistency check due to order reduction in root solving operation in the case of 5th order
                         if(tau.gt.tau_steps_list(i_step_root - 1)) then
+print *, 'Warning: tau', tau, 'tau_steps_list',tau_steps_list(i_step_root - 1)
                             print *, 'Warning: Final t_hamiltonian step in 0th order for avoiding inconsistency.'
                             tau = t_remain_new/dt_dtau_const
                         endif
