### 2022-10-04 172c09a
Working on Hamiltonian time
