+++ SRC/pusher_tetra_poly.f90
@@ -537,6 +537,7 @@ if(diag_pusher_tetry_poly) print *, 't_pass',t_pass
 print *, 'Warning: tau', tau, 'tau_steps_list',tau_steps_list(i_step_root - 1)
 print *, 'number_of_integration_steps', number_of_integration_steps
 print *, 'tau_steps_list',tau_steps_list
+print *, 'tau_steps_list * dt_dtau',tau_steps_list*dt_dtau_const
 print *, 't_hamiltonian_list',t_hamiltonian_list
 print *, 't_remain',t_remain
 print *, 't_remain_new',t_remain_new
