### 2022-10-04 1d41577
Working on Hamiltonian time
