### 2022-10-04 ef08c9d
Working on Hamiltonian time
