### 2022-10-04 2682542
Working on Hamiltonian time
