### 2022-10-04 153465a
Working on Hamiltonian time
