### 2022-10-05 82449f8
Working on negative sqrtg in rk4 pusher
