+++ SRC/pusher_tetra_poly.f90
@@ -20,6 +20,11 @@ module pusher_tetra_poly_mod
 
     implicit none
 !
+    !change those for adaptive step sizes, probably allocatable
+    double precision, dimension(:), allocatable, public, protected            :: tau_steps_list
+    double precision, dimension(:,:), allocatable, public, protected          :: intermediate_z0_list
+    integer, public, protected                                                :: number_of_integration_steps
+!
     private
 !    
     integer                             :: iface_init
@@ -37,11 +42,6 @@ module pusher_tetra_poly_mod
                                0.d0, 0.d0, 1.d0, 0.d0, &
                                0.d0, 0.d0, 0.d0, 1.d0 ], [4,4])
 !
-    !change those for adaptive step sizes, probably allocatable
-    double precision, dimension(:), allocatable           :: tau_steps_list
-    double precision, dimension(:,:), allocatable         :: intermediate_z0_list
-    integer                                               :: number_of_integration_steps
-!
     !Diagnostics for adaptive step scheme (only useable for one particle calculation)
     double precision, dimension(:,:), allocatable         :: total_fluctuation_report, single_step_fluctuation_report, &
                                                            & closure_fluctuation_report
@@ -53,7 +53,8 @@ module pusher_tetra_poly_mod
     !$OMP& z_init,k1,k3,vmod0,tau_steps_list,intermediate_z0_list,number_of_integration_steps,sign_rhs)
 !
     public :: pusher_tetra_poly,initialize_const_motion_poly, &
-        & Quadratic_Solver2, Cubic_Solver, Quartic_Solver,analytic_integration_without_precomp,energy_tot_func
+        & Quadratic_Solver2, Cubic_Solver, Quartic_Solver,energy_tot_func, &
+        & set_integration_coef_manually
 !   
     contains
 !
@@ -3076,9 +3077,12 @@ module par_adiab_inv_poly_mod
         !           the quantities from last pushing
 !        
         use poly_without_precomp_mod, only: amat,b
-        use pusher_tetra_poly_mod, only: dt_dtau_const,z_init,ind_tetr,analytic_integration_without_precomp, &
-                                                & energy_tot_func
+        use pusher_tetra_poly_mod, only: dt_dtau_const,z_init,ind_tetr, &
+                                                & energy_tot_func, &
+                                                number_of_integration_steps, intermediate_z0_list, tau_steps_list, &
+                                                & set_integration_coef_manually
         use tetra_physics_mod, only: particle_mass,tetra_physics
+        use gorilla_diag_mod, only: diag_pusher_tetry_poly_adaptive
 !  
         implicit none
 !
@@ -3088,6 +3092,10 @@ module par_adiab_inv_poly_mod
         double precision                    :: tau,a44,b4,tau_part1
         double precision, dimension(4)      :: z
         double precision, dimension(3)      :: x
+        integer                             :: i, turning_index
+!
+        !If particle was lost and the recording quantities deallocated -> no further calculations (here for safety)
+        if (.not.allocated(intermediate_z0_list)) return
 !
         nskip = n_skip_vpar_0
 !
@@ -3098,23 +3106,53 @@ module par_adiab_inv_poly_mod
         a44 = amat(4,4)
         b4 = b(4)
 !
+if(diag_pusher_tetry_poly_adaptive) then
+if(number_of_integration_steps.gt.1) then
+print*, '----------------------------------'
+do i = 1, number_of_integration_steps
+print*, 'z0 number', i
+print*, intermediate_z0_list(:,i)
+print*, tau_steps_list(i)
+enddo
[…]
