+++ SRC/gorilla_plot_mod.f90
@@ -397,11 +397,19 @@ module gorilla_plot_mod
 !
         subroutine load_gorilla_plot_inp()
 !
+            use gorilla_settings_mod, only: i_time_tracing_option
+!
             open(unit=90, file='gorilla_plot.inp', status='unknown')
             read(90,nml=gorilla_plot_nml)
             close(90)
 !
             print *,'GORILLA Plotting: Loaded input data from gorilla_plot.inp'
+!
+            if(boole_J_par) then
+                if(i_time_tracing_option.eq.2) print *, 'WARNING: Time evolution for J_par is computed in zeroth order.' &
+                                                        & 'Whereas, Time evolution of orbit is computed exactly.' &
+                                                        & 'For consistency, set i_time_tracing_option to 1.'
+            endif
 !
         end subroutine load_gorilla_plot_inp
 !
