### 2022-10-10 93d20aa
grid_for_find_tetra correctly implemented
