+++ SRC/find_tetra_mod.f90
@@ -1,0 +1,517 @@
+!
+!ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
+!
+module find_tetra_mod
+!
+    implicit none
+!
+    private
+!
+    public :: grid_for_find_tetra, find_tetra
+!
+    integer, dimension(:,:), allocatable           :: equidistant_grid
+    integer, dimension(:), allocatable             :: entry_counter
+    logical                                        :: boole_axi_symmetry
+    integer                                        :: a,b,c, a_factor, b_factor, c_factor,na, nb, nc, max_nb
+    integer                                        :: ind_a, ind_b, ind_c
+    double precision                               :: amin, amax, cmin, cmax, bmin, bmax, delta_a, delta_b, delta_c
+    double precision, dimension(:,:), allocatable  :: box_centres, save_box_centres
+!
+    contains
+!
+!ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
+!
+    subroutine grid_for_find_tetra
+!
+        use constants, only: pi, eps
+        use tetra_grid_mod, only: verts_rphiz, verts_sthetaphi, tetra_grid, ntetr
+        use tetra_grid_settings_mod, only: grid_size, grid_kind
+        use tetra_physics_mod, only: tetra_physics, coord_system
+        !use pusher_tetra_poly_mod, only: normal_distance_func
+!
+        implicit none            
+        double precision                               :: tetr_amin, tetr_amax, tetr_cmin, tetr_cmax, half_diagonal
+        integer                                        :: i,k,m, box_index, num_columns, num_hexahedra, maxtetr, index_change
+        integer, dimension(2)                          :: steps_a, steps_b, steps_c
+        logical                                        :: boole_reduce_entries
+        double precision, dimension(4)                 :: normal_distances, normalisations
+        double precision, dimension(3)                 :: current_box_centre
+        double precision, dimension(:,:), allocatable  :: verts_abc
+!
+print*, 'grid_for_find_tetra started'
+        a_factor = 2
+        c_factor = 2
+        boole_reduce_entries = .true.
+        boole_axi_symmetry = .false.
+!
+        if (coord_system.eq.1) allocate(verts_abc(size(verts_rphiz(:,1)),size(verts_rphiz(1,:))))
+        if (coord_system.eq.2) allocate(verts_abc(size(verts_sthetaphi(:,1)),size(verts_sthetaphi(1,:))))
+!
+        if (coord_system.eq.1) verts_abc = verts_rphiz
+        if (coord_system.eq.2) verts_abc = verts_sthetaphi
+!
+        ind_a = 1 !(R in cylindrical coordinates, s in flux coordinates)
+        ind_b = 2 !(phi in cylindrical and flux coordinates)
+        ind_c = 3 !(z in cylindrical coordinates, theta in flux coordinates)
+!
+        if (coord_system.eq.2) then
+            ind_b = 3
+            ind_c = 2
+        endif
+!
+        amin = minval(verts_abc(ind_a,:)) - 2*eps
+        amax = maxval(verts_abc(ind_a,:)) + 2*eps
+        bmin = 0
+        bmax = 2*pi
+        cmin = minval(verts_abc(ind_c,:)) - 2*eps
+        cmax = maxval(verts_abc(ind_c,:)) + 2*eps
+        if (boole_axi_symmetry) bmax = 2*pi/grid_size(ind_b)
+!
+        na = grid_size(ind_a)*a_factor
+        nc = grid_size(ind_c)*c_factor
+!
+        if (grid_kind.eq.4) then
+            na = 200*a_factor
+            nc = 200*c_factor
[…]
