### 2022-10-11 #4576 @krystophny
> @krystophny, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Thanks for the request! For this submission, I am sorry that I cannot act as a reviewer, since I am a co-author (not immediately obvious since there's no direct link between my GitHub user and real name).
