### 2022-10-14 006658b
sign_t_step_save is now set to thread private
