### 2022-10-20 14976af
Adjust examples to account correct time dynamics (VMEC)
