+++ .github/workflows/Mac.yml
@@ -13,17 +13,21 @@ jobs:
     runs-on: macos-latest
     
     env:
-      FC: gfortran-9
+      CC: gcc
+      CXX: gcc
+      FC: gfortran
       PFUNIT_DIR: /Users/runner/work/GORILLA/GORILLA/pFUnit/build/installed/PFUNIT-4.4
 
     steps:
       - name: Checkout
-        uses: actions/checkout@v2
+        uses: actions/checkout@v3
 
       - name: Dependencies
         run: |
           curl -LO https://raw.githubusercontent.com/GiovanniBussi/macports-ci/master/macports-ci
           source ./macports-ci install
+          sudo port install gcc11 +gfortran
+          sudo port select --set gcc mp-gcc11
           sudo port install netcdf netcdf-fortran
           sudo port install lcov
           brew install libomp
@@ -51,7 +55,7 @@ jobs:
           ./build_coverage.sh
 
       - name: Archive code coverage results
-        uses: actions/upload-artifact@v2
+        uses: actions/upload-artifact@v3
         with:
           name: code-coverage-report
           path: BUILD/COVERAGE/
@@ -61,17 +65,21 @@ jobs:
     runs-on: macos-latest
     
     env:
-      FC: gfortran-9
+      CC: gcc
+      CXX: gcc
+      FC: gfortran
       PFUNIT_DIR: /home/runner/work/GORILLA/GORILLA/pFUnit/build/installed/PFUNIT-4.4
 
     steps:
       - name: Checkout
-        uses: actions/checkout@v2
+        uses: actions/checkout@v3
 
       - name: Dependencies
         run: |
           curl -LO https://raw.githubusercontent.com/GiovanniBussi/macports-ci/master/macports-ci
           source ./macports-ci install
+          sudo port install gcc11 +gfortran
+          sudo port select --set gcc mp-gcc11
           sudo port install netcdf netcdf-fortran
           
       - name: Additional Files
