### 2022-10-24 d2a6609
Moment calculation implemented
