+++ SRC/orbit_timestep_gorilla.f90
@@ -7,7 +7,7 @@ module orbit_timestep_gorilla_mod
 !
     private
 !
-    public :: orbit_timestep_gorilla,initialize_gorilla,check_coordinate_domain,find_tetra
+    public :: orbit_timestep_gorilla,initialize_gorilla,check_coordinate_domain
 !
     integer, dimension(:,:), allocatable, public, protected   :: equidistant_grid
     integer, dimension(:), allocatable, public, protected     :: entry_counter
@@ -16,7 +16,7 @@ module orbit_timestep_gorilla_mod
 !    
     contains
 !
-        subroutine orbit_timestep_gorilla(x,vpar,vperp,t_step,boole_initialized,ind_tetr,iface,t_remain_out)!,boole_grid_for_find_tetra)
+        subroutine orbit_timestep_gorilla(x,vpar,vperp,t_step,boole_initialized,ind_tetr,iface,t_remain_out)
 !
             use supporting_functions_mod, only: bmod_func, vperp_func
             use pusher_tetra_rk_mod, only: pusher_tetra_rk,initialize_const_motion_rk
@@ -26,6 +26,7 @@ module orbit_timestep_gorilla_mod
             use tetra_physics_mod, only: tetra_physics,particle_charge,particle_mass
             !use tetra_grid_mod, only: ntetr
             use gorilla_settings_mod, only: ipusher, poly_order
+            use find_tetra_mod, only: find_tetra
 !
             implicit none
 !
@@ -35,7 +36,6 @@ module orbit_timestep_gorilla_mod
             logical, intent(inout)                          :: boole_initialized
             integer, intent(inout)                          :: ind_tetr,iface
             double precision, intent(out), optional         :: t_remain_out
-            !logical, intent(in), optional                   :: boole_grid_for_find_tetra
             double precision, dimension(3)                  :: z_save
             double precision                                :: vperp2,t_remain,t_pass,vpar_save
             logical                                         :: boole_t_finished
@@ -147,19 +147,21 @@ module orbit_timestep_gorilla_mod
 !
 !ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
 !
-        subroutine initialize_gorilla(boole_grid_for_find_tetra,i_option,ipert_in,bmod_multiplier)
+        subroutine initialize_gorilla(i_option,ipert_in,bmod_multiplier)!,boole_grid_for_find_tetra)
 !
             use constants, only: echarge,ame,amp,clight
-            use gorilla_settings_mod, only: eps_Phi,coord_system,ispecies
+            use gorilla_settings_mod, only: eps_Phi,coord_system,ispecies,boole_grid_for_find_tetra
             use tetra_grid_mod, only: make_tetra_grid
             use tetra_physics_mod, only: make_tetra_physics,check_tetra_overlaps,cm_over_e,particle_charge,particle_mass
             use tetra_physics_poly_precomp_mod, only: make_precomp_poly
             use tetra_grid_settings_mod, only: grid_kind
+            use find_tetra_mod, only: grid_for_find_tetra
+
 !
             implicit none
 !
             integer                                 :: iper,ipert
-            logical, intent(inout), optional        :: boole_grid_for_find_tetra
+            !logical, intent(inout), optional        :: boole_grid_for_find_tetra
             integer,intent(in),optional             :: i_option,ipert_in
             logical                                 :: boole_grid,boole_physics,boole_bmod_multiplier
             double precision,intent(in),optional    :: bmod_multiplier
@@ -261,501 +263,14 @@ module orbit_timestep_gorilla_mod
 !
                 call make_precomp_poly()
 !
[…]
