+++ SRC/pusher_tetra_poly_version _without_interface.f90
@@ -20,63 +20,10 @@ module pusher_tetra_poly_mod
 
     implicit none
 !
-    interface poly_multiplication
-    pure function scalar_multiplication_without_precomp(scalar_coef_1,scalar_coef_2)
-        implicit none
-        double precision, dimension(:), intent(in)           :: scalar_coef_1,scalar_coef_2
-        double precision, dimension(:), allocatable          :: scalar_multiplication_without_precomp
-    end function scalar_multiplication_without_precomp
-!
-    pure function vector_multiplication_without_precomp(scalar_coef,vector_coef)
-        implicit none
-        double precision, dimension(:), intent(in)           :: scalar_coef
-        double precision, dimension(:,:), intent(in)         :: vector_coef
-        double precision, dimension(:,:), allocatable        :: vector_multiplication_without_precomp
-        end function vector_multiplication_without_precomp
-!
-    pure function tensor_multiplication_without_precomp(vector_coef_1,vector_coef_2)
-        implicit none
-        double precision, dimension(:,:), intent(in)         :: vector_coef_1,vector_coef_2
-        double precision, dimension(:,:,:), allocatable      :: tensor_multiplication_without_precomp
-        end function tensor_multiplication_without_precomp
-    end interface poly_multiplication
-!
-    interface moment_integration
-    pure function scalar_integral_without_precomp(poly_order,tau,scalar_coef)
-        implicit none
-        integer, intent(in)                                  :: poly_order
-        double precision, intent(in)                         :: tau
-        double precision, dimension(:), intent(in)           :: scalar_coef
-        double precision                                     :: scalar_integral_without_precomp
-    end function scalar_integral_without_precomp
-!
-    pure function vector_integral_without_precomp(poly_order,tau,vector_coef)
-        implicit none
-        integer, intent(in)                                   :: poly_order
-        double precision, intent(in)                          :: tau
-        double precision, dimension(:,:), intent(in)          :: vector_coef
-        double precision, dimension(:), allocatable           :: vector_integral_without_precomp
-    end function vector_integral_without_precomp
-!
-    pure function tensor_integral_without_precomp(poly_order,tau,tensor_coef)
-        implicit none
-        integer, intent(in)                             :: poly_order
-        double precision, intent(in)                    :: tau
-        double precision, dimension(:,:,:), intent(in)  :: tensor_coef
-        double precision, dimension(:,:), allocatable   :: tensor_integral_without_precomp
-    end function tensor_integral_without_precomp
-    end interface moment_integration
-!
-    !change those for adaptive step sizes, probably allocatable
-    double precision, dimension(:), allocatable, public, protected            :: tau_steps_list
-    double precision, dimension(:,:), allocatable, public, protected          :: intermediate_z0_list
-    integer, public, protected                                                :: number_of_integration_steps
-!
     private
 !    
     integer                             :: iface_init
     integer, public, protected          :: ind_tetr
-    integer, public, protected          :: sign_rhs
     double precision                    :: perpinv2,vmod0
     double precision, public, protected :: perpinv,dt_dtau_const,bmod0
     double precision                    :: t_remain
[…]
