+++ SRC/find_tetra_mod.f90
@@ -12,7 +12,7 @@ module find_tetra_mod
     integer, dimension(:,:), allocatable           :: equidistant_grid
     integer, dimension(:), allocatable             :: entry_counter
     logical                                        :: boole_axi_symmetry
-    integer                                        :: a,b,c, a_factor, b_factor, c_factor,na, nb, nc, max_nb
+    integer                                        :: a,b,c, b_factor, c_factor,na, nb, nc, max_nb
     integer                                        :: ind_a, ind_b, ind_c
     double precision                               :: amin, amax, cmin, cmax, bmin, bmax, delta_a, delta_b, delta_c
     double precision, dimension(:,:), allocatable  :: box_centres, save_box_centres
@@ -27,6 +27,7 @@ module find_tetra_mod
         use tetra_grid_mod, only: verts_rphiz, verts_sthetaphi, tetra_grid, ntetr
         use tetra_grid_settings_mod, only: grid_size, grid_kind, n_field_periods
         use tetra_physics_mod, only: tetra_physics, coord_system
+        use gorilla_settings_mod, only: a_factor
         !use pusher_tetra_poly_mod, only: normal_distance_func
 !
         implicit none            
@@ -39,7 +40,6 @@ module find_tetra_mod
         double precision, dimension(:,:), allocatable  :: verts_abc
 !
 print*, 'grid_for_find_tetra started'
-        a_factor = 4
         boole_reduce_entries = .true.
         boole_axi_symmetry = .false.
 !
@@ -206,6 +206,7 @@ print*, 'abc_factor, delta_abc = ', a_factor, b_factor, c_factor, delta_a, delta
 !
  PRINT*, 'average number of entries before reduction is: ', sum(entry_counter), 'divided by', num_hexahedra, ' = ', &
  dble(sum(entry_counter))/dble(num_hexahedra)
+ print*, 'entry counter before reduction is ', entry_counter(14129)
 !
         if (boole_reduce_entries) then
             do a = 1,na!fill 1st column (R in cylindrical coordinates, s in flux coordinates)
@@ -245,24 +246,14 @@ print*, 'abc_factor, delta_abc = ', a_factor, b_factor, c_factor, delta_a, delta
                     normal_distances(1) = normal_distances(1) + &
                                           tetra_physics(equidistant_grid(i,k+index_change))%dist_ref/normalisations(1)
                     if (any(normal_distances.lt.-half_diagonal)) then
-! if ((equidistant_grid(i,k+index_change).eq.94857).and.(i.eq.255634)) then
-!     print*, 'got xou', normal_distances, half_diagonal
-!     print*, verts_abc(ind_a,tetra_grid(94857)%ind_knot([1,2,3,4]))
-!     print*, verts_abc(ind_b,tetra_grid(94857)%ind_knot([1,2,3,4]))
-!     print*, verts_abc(ind_c,tetra_grid(94857)%ind_knot([1,2,3,4]))
-!     print*, m
-!     print*, tetra_physics(equidistant_grid(i,k+index_change))%anorm(:,4)/normalisations(4)
-!     print*, tetra_physics(equidistant_grid(i,k+index_change))%anorm(:,4)
-!     print*, normalisations
-!     print*, normal_distances
-!     print*, tetra_physics(equidistant_grid(i,k+index_change))%dist_ref
-!     print*, current_box_centre
-!     print*, tetra_physics(equidistant_grid(i,k+index_change))%x1
-!     print*, box_centres(i,:)
-! endif
[…]
