+++ Makefile
@@ -50,6 +50,7 @@ SOURCES = SetWorkingPrecision.f90\
 	pusher_tetra_func_mod.f90 \
 	pusher_tetra_poly.f90 \
 	pusher_tetra_rk.f90 \
+	find_tetra_mod.f90 \
 	get_canonical_coordinates.f90 \
 	orbit_timestep_gorilla.f90 \
 	gorilla_plot_mod.f90 \
