### 2022-10-27 #4576 @pibion
@jordigh, @nicoguaro, @moorepants, @IgorBaratta, @mzamoraz, @shibabrat, @dawbarton, @EduPH, @inakleinbottle, @joglekara, @pratikvn, @hasantahir, @joshia5, @rogeriojorge, @akaptano, @yhtang, @dl11g11, @hasantahir -

would any of you be interested in reviewing this submission for JOSS? Reviewing will mean going though a checklist to make sure the code meets JOSS publication standards (for example, there's instructions on installation, a test suite, and contribution instructions).
