### 2022-10-28 #4890 @editorialbot
[REVIEW]: GORILLA: Guiding-center ORbit Integration with Local Linearization Approach
**Submitting author:** @micheder (<a href="http://orcid.org/0000-0002-1392-1664">Michael Eder</a>)
**Repository:** https://github.com/itpplasma/GORILLA
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.2
