### 2022-11-03 #4576 @danielskatz
@Shibabrat the error is misleading here - the problem is actually that you are doing this in the pre-review issue, not the review issue - try the command again in https://github.com/openjournals/joss-reviews/issues/4890 please
