### 2022-11-23 9be75ee
Some additional diagnostics for Soledge3X-EIRENE mesh
