+++ SRC/gorilla_diag_mod.f90
@@ -5,6 +5,7 @@
     logical :: report_pusher_tetry_poly_adaptive = .false.
     logical :: diag_pusher_tetra_rk = .false.
     logical :: transp_coef_trigger = .false.
+    logical :: diag_mesh_SOLEDGE3X_EIRENE = .false.
   end module gorilla_diag_mod
 !
 !ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
