+++ SRC/find_tetra_mod.f90
@@ -12,7 +12,7 @@ module find_tetra_mod
     integer, dimension(:,:), allocatable           :: equidistant_grid
     integer, dimension(:), allocatable             :: entry_counter
     logical                                        :: boole_axi_symmetry
-    integer                                        :: a,b,c, b_factor, c_factor,na, nb, nc, max_nb
+    integer                                        :: a,b,c,na, nb, nc, max_nb
     integer                                        :: ind_a, ind_b, ind_c
     double precision                               :: amin, amax, cmin, cmax, bmin, bmax, delta_a, delta_b, delta_c
     double precision, dimension(:,:), allocatable  :: box_centres, save_box_centres
@@ -27,7 +27,7 @@ module find_tetra_mod
         use tetra_grid_mod, only: verts_rphiz, verts_sthetaphi, tetra_grid, ntetr
         use tetra_grid_settings_mod, only: grid_size, grid_kind, n_field_periods
         use tetra_physics_mod, only: tetra_physics, coord_system
-        use gorilla_settings_mod, only: a_factor
+        use gorilla_settings_mod, only: a_factor, b_factor, c_factor
         !use pusher_tetra_poly_mod, only: normal_distance_func
 !
         implicit none            
@@ -38,7 +38,9 @@ module find_tetra_mod
         double precision, dimension(4)                 :: normal_distances, normalisations
         double precision, dimension(3)                 :: current_box_centre
         double precision, dimension(:,:), allocatable  :: verts_abc
+        integer :: t1, t2, clock_max, clock_rate1, clock_rate2
 !
+call system_clock (t1, clock_rate1, clock_max)
 print*, 'grid_for_find_tetra started'
         boole_reduce_entries = .true.
         boole_axi_symmetry = .false.
@@ -73,7 +75,7 @@ print*, 'grid_for_find_tetra started'
         na = grid_size(ind_a)*a_factor
         delta_a = (amax - amin)/na
 !
-        c_factor = maxval((/nint((cmax-cmin)/(grid_size(ind_c)*delta_a)),1/))
+        if (c_factor.eq.0) c_factor = maxval((/nint((cmax-cmin)/(grid_size(ind_c)*delta_a)),1/))
         nc = grid_size(ind_c)*c_factor
 !
         if (grid_kind.eq.4) then
@@ -87,7 +89,7 @@ print*, 'grid_for_find_tetra started'
 !
         delta_c = (cmax - cmin)/nc
 !
-        b_factor = maxval((/nint(2*pi/(grid_size(ind_b)*n_field_periods*((delta_a+delta_c)/2))),1/))
+        if (b_factor.eq.0) b_factor = maxval((/nint(2*pi/(grid_size(ind_b)*n_field_periods*sqrt((delta_a**2+delta_c**2)/2))),1/))
         b_factor = b_factor*n_field_periods
 !
         nb = grid_size(ind_b)*b_factor
@@ -97,6 +99,7 @@ print*, 'abc_factor, delta_abc = ', a_factor, b_factor, c_factor, delta_a, delta
         max_nb = nb
         if (boole_axi_symmetry) max_nb = b_factor
         num_hexahedra = na*max_nb*nc
+!print*, 'num_hexahedra = ', num_hexahedra, na, max_nb, nc, na*nc, na*max_nb, max_nb*nc, na*max_nb*nc
 !
         allocate(entry_counter(num_hexahedra))
         allocate(box_centres(num_hexahedra,3))
@@ -206,7 +209,11 @@ print*, 'abc_factor, delta_abc = ', a_factor, b_factor, c_factor, delta_a, delta
 !
[…]
