### 2022-12-11 #4890 @rogeriojorge
Dear authors of GORILLA, here is my review:
--------------------------------------------
MANUSCRIPT
The introduction is extremely well written and does a good job motivating particle tracing as a whole. However, it fails to motivate the need for another particle tracer code that simulates the scrape-off layer region of fusion devices. For example, the BEAMS3D code is able to trace particles through the separatrix https://iopscience.iop.org/article/10.1088/0741-3335/56/9/095019, as well as the RGCO code https://catalog.lib.kyushu-u.ac.jp/opac_download_md/5909096/2022_p228.pdf . Is it the fact that instead of providing coil shapes, this code uses g-files provided by the SOLEDGE3X-EIRENE framework? Furthermore, it uses different integration algorithms which can add to its novelty. However, this should be mentioned in the introduction.
In the summary, tI was expecting to see something about the inputs and outputs of the code. While the figure is helpful, one of the main features, the Poincaré cut, is not clearly explaind in the main text and is actually one of the main outputs of the code.
In the statement of need, can the authors add the possibility of using such framework in a single-stage stellarator optimization routine where the VMEC and coil shapes are varied simultaneously?

CODE
 Installation: the make command is not successful to install (at least not in MacOS with the makefile provided) but such feature is mentioned in the README. Nevertheless, the build.sh command works. Remark: the README mentions build.x and build_coverage.x but the extension is actually .sh
  Unit tests: please provide a simple way to check what are the tests performed and the overall code coverage of the code.
