### 2023-01-06 #4890 @hasantahir
The repository along with the submission is extremely well written. My take on `GORILLA` is that it is free from any topology limitations which enhances its usefulness over existing I solvers.

I have been able to build and run the code on a macOS machine. Also tried the Matlab `.m` files which work.
