+++ .github/workflows/Ubuntu.yml
@@ -14,7 +14,7 @@ jobs:
     
     env:
       FC: gfortran
-      PFUNIT_DIR: /home/runner/work/GORILLA/GORILLA/pFUnit/build/installed/PFUNIT-4.4/
+      PFUNIT_DIR: /home/runner/work/GORILLA/GORILLA/pFUnit/build/installed/PFUNIT-4.6/
 
     steps:
       - name: Checkout
@@ -59,7 +59,7 @@ jobs:
     
     env:
       FC: gfortran
-      PFUNIT_DIR: /home/runner/work/GORILLA/GORILLA/pFUnit/build/installed/PFUNIT-4.4
+      PFUNIT_DIR: /home/runner/work/GORILLA/GORILLA/pFUnit/build/installed/PFUNIT-4.6
 
     steps:
       - name: Checkout
