### 2023-01-16 4ce9734
Started framework for getting the E-field
