+++ Makefile
@@ -41,6 +41,7 @@ SOURCES = SetWorkingPrecision.f90\
 	tetra_grid_mod.f90 \
 	make_grid_rect.f90 \
 	bdivfree.f90 \
+	strong_electric_field_mod.f90 \
 	tetra_physics_mod.f90 \
 	tetra_physics_poly_precomp_mod.f90 \
 	differentiate.f90 \
