### 2023-01-18 #4890 @GeorgGrassler
Dear @rogeriojorge,

sorry for coming back to your reply so late, but there was a change of jurisdiction over the repository at our institute. Therefore, there are some new people looking over the development of GORILLA, including me. We will look into your suggestions, but could you elaborate a bit about the issue with the Unit tests?

> Unit tests: please provide a simple way to check what are the tests performed

Do you mean, which of the unit tests are executed by the coverage? That should be displayed in the console as build_coverage.sh is executed (around the end of the log), as well as a brief overview of their results. Or is it about the content of what is actually tested i.e. a short description should be additional given out in the log?

>and the overall code coverage of the code.
The report should be in the BUILD/COVERAGE folder as the index.html file, as explained in the README. Is it the placement of the report that is the issue? We thought that following the standard protocol of ctest in terms of storing the report in the automatically generated COVERAGE folder would be appropriate.

We would be grateful about any input you could give us, so that we can make the correct changes.
