### 2023-01-19 #4890 @rogeriojorge
Dear @GeorgGrassler

No worries.

1. My comment regarding tests comes from the fact that the folder `TESTS` inside the `SRC` folder is poorly documented. It should be mentioned in the README that the unit tests performed by GORILLA are in this folder (if this is true) and what is the policy regarding testing. From what I can find, besides the name, there is little way of knowing what tests are being performed. This can be fixed by adding a README file in the `TESTS` folder or by including comments at the beginning of each file stating its purpose. There should also be a policy regarding testing stated in the main README or documentation in order for tests to be coherent. For example, in the file `test_pusher_tetra_rk.f90` the actual assertion is commented out, leaving the test to only an initialization check.

2. Thank you for clarifying where to find the code coverage. That is fine with me. Although is completely optional, it might be a good idea to display the coverage directly on the repo (as a badge, for example) without the need for the user to download the files and create the report him\herself.
