### 2023-01-24 5701aa2
Update build_coverage.sh
