### 2023-02-07 fa25cbc
Fixed CMake and Make for macos arm64
