+++ CMakeLists.txt
@@ -2,13 +2,14 @@ cmake_minimum_required(VERSION 3.12)
 
 project (GORILLA
   VERSION 1.0.0)
-  
+
 enable_language(Fortran)
 
+include_directories(/usr/local/include)
 if(UNIX AND NOT APPLE)
 	include_directories(/usr/include)
-elseif(APPLE)
-	include_directories(/opt/local/include)
+elseif(APPLE AND ${CMAKE_HOST_SYSTEM_PROCESSOR} STREQUAL "arm64")
+	include_directories(/opt/homebrew/include)
 endif()
 
 find_package(BLAS REQUIRED)
@@ -23,8 +24,9 @@ endif()
 add_compile_options(-g -fbacktrace -ffpe-trap=zero,overflow,invalid -fbounds-check -fopenmp)
 add_link_options(-g -fbacktrace -ffpe-trap=zero,overflow,invalid -fbounds-check -fopenmp)
 
-if(APPLE)
-	add_link_options(-L/opt/local/lib)
+add_link_options(-L/usr/local/lib)
+if(APPLE AND ${CMAKE_HOST_SYSTEM_PROCESSOR} STREQUAL "arm64")
+	add_link_options(-L/opt/homebrew/lib)
 endif()
 
 if ("$ENV{GORILLA_COVERAGE}" STREQUAL "TRUE")
@@ -33,3 +35,4 @@ endif()
 
 add_subdirectory(SRC)
 
+message(${CMAKE_HOST_SYSTEM_PROCESSOR})
