+++ SRC/strong_electric_field_mod.f90
@@ -5,11 +5,17 @@ module strong_electric_field_mod
     private
 !
     !Setting mode of getting the electric potential
-    integer, parameter                           :: electric_potential_mode = 1
+    !2 ... potential as a rescaled poloidal flux (E-field orthogonal to magnetic surfaces)
+    integer, parameter                           :: electric_potential_option = 2
     !Setting mode of getting the electric field
-    integer, parameter                           :: electric_field_mode = 1
+    !1 ... using central differences to estimate the gradient of a scalar potential
+    integer, parameter                           :: electric_field_option = 1
 !
-    public :: get_electric_field
+    !Stepsize for central difference approach (only calculated at initial call)
+    logical                                      :: boole_first_call = .true.
+    double precision, dimension(3)               :: dx_array
+!
+    public :: get_electric_field, save_electric_field
 !
 contains
 !
@@ -22,30 +28,14 @@ contains
         double precision, intent(in)             :: R, phi, Z
         double precision, intent(out)            :: E_x1, E_x2, E_x3
 !
-        double precision, dimension(6)           :: electric_potential
-        double precision, dimension(3)           :: dx_array, dx
-        integer                                  :: i
+        select case(electric_field_option)
 !
-        select case(electric_field_mode)
-!
-            !Get electric field from scalar electric potential
+            !Get electric field from scalar electric potential via central difference
             case(1)
-                do i = 1,3
-                    dx = 0.0d0
-                    !Forward step
-                    dx(i) = dx_array(i)
-                    call get_electric_potential(R + dx(1), phi + dx(1), Z + dx(1), electric_potential(2*i-1))
-                    !Backward step
-                    dx = -dx
-                    call get_electric_potential(R + dx(1), phi + dx(1), Z + dx(1), electric_potential(2*i))
-                end do
-                !Forming the central difference as approximation of derivative
-                E_x1 = -(electric_potential(1)-electric_potential(2))/(2*dx_array(1))
-                E_x2 = -(electric_potential(3)-electric_potential(4))/(2*dx_array(2))
-                E_x3 = -(electric_potential(5)-electric_potential(6))/(2*dx_array(3))
+                call electric_field_from_central_difference(R,phi,Z,E_x1,E_x2,E_x3)
 !
             case default
-                print *, 'ERROR: Invalid mode to get electric field. Change the parameters in module strong_electric_field_mod!'
+                print *, 'ERROR: Invalid option to get electric field. Change the parameters in module strong_electric_field_mod!'
                 stop
         end select
 !
@@ -53,24 +43,162 @@ contains
 !
 !ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
 !
+    subroutine electric_field_from_central_difference(R,phi,Z,E_x1,E_x2,E_x3)
+!
+        implicit none
+!
+        double precision, intent(in)             :: R, phi, Z
+        double precision, intent(out)            :: E_x1, E_x2, E_x3
+!
+        double precision, dimension(6)           :: electric_potential
+        double precision, dimension(3)           :: dx
+        integer                                  :: i
+        double precision, parameter              :: machine_epsilon_3 = 1d-6
+!
+        !Determine the stepsize
+        if(boole_first_call) then
[…]
