+++ SRC/pusher_tetra_poly.f90
@@ -46,7 +46,7 @@ module pusher_tetra_poly_mod
     !$OMP& z_init,k1,k3,vmod0,tau_steps_list,intermediate_z0_list,number_of_integration_steps,sign_rhs)
 !
     public :: pusher_tetra_poly,initialize_const_motion_poly, &
-        & Quadratic_Solver2, Cubic_Solver, Quartic_Solver,energy_tot_func, &
+        & Quadratic_Solver2, Cubic_Solver, Quartic_Solver, &
         & set_integration_coef_manually
 !   
     contains
@@ -110,6 +110,7 @@ endif
         subroutine initialize_pusher_tetra_poly(ind_tetr_inout,x,iface,vpar,t_remain_in)
 !
             use tetra_physics_mod, only: tetra_physics, sign_sqg
+            use supporting_functions_mod, only: bmod_func, phi_elec_func
 !
             implicit none
 !
@@ -141,10 +142,10 @@ endif
             dt_dtau_const = dt_dtau_const*dble(sign_rhs)
 !    
             !Module of B at the entry point of the particle
-            bmod0 = bmod_func(z_init(1:3)) !tetra_physics(ind_tetr)%bmod1+sum(tetra_physics(ind_tetr)%gb*z_init(1:3))
+            bmod0 = bmod_func(z_init(1:3),ind_tetr) !tetra_physics(ind_tetr)%bmod1+sum(tetra_physics(ind_tetr)%gb*z_init(1:3))
 !
             !Phi at the entry point of the particle
-            phi_elec = phi_elec_func(z_init(1:3))   !tetra_physics(ind_tetr)%Phi1+sum(tetra_physics(ind_tetr)%gPhi*z_init(1:3))
+            phi_elec = phi_elec_func(z_init(1:3),ind_tetr)   !tetra_physics(ind_tetr)%Phi1+sum(tetra_physics(ind_tetr)%gPhi*z_init(1:3))
 !
             !Auxiliary quantities
             vperp2 = -2.d0*perpinv*bmod0
@@ -805,6 +806,7 @@ if(diag_pusher_tetry_poly)  print *, 'Error: Tau is out of safety boundary'
 !
         use gorilla_settings_mod, only: desired_delta_energy, max_n_intermediate_steps
         use gorilla_diag_mod, only: diag_pusher_tetry_poly, report_pusher_tetry_poly_adaptive
+        use supporting_functions_mod, only: energy_tot_func
 !
         implicit none
 !
@@ -846,8 +848,8 @@ if(diag_pusher_tetry_poly) print*, 'z', z
         if (.not.boole_face_correct) return
 !
         z0 = intermediate_z0_list(:,number_of_integration_steps)
-        energy_start = energy_tot_func(z0)
-        energy_current = energy_tot_func(z)
+        energy_start = energy_tot_func(z0,perpinv,ind_tetr)
+        energy_current = energy_tot_func(z,perpinv,ind_tetr)
         delta_energy_current = abs(1-energy_current/energy_start)
 !
 if(diag_pusher_tetry_poly) print*, 'z0', z0
@@ -888,6 +890,7 @@ endif
 !
         use gorilla_settings_mod, only: i_precomp, desired_delta_energy, max_n_intermediate_steps
         use gorilla_diag_mod,only: diag_pusher_tetry_poly_adaptive, report_pusher_tetry_poly_adaptive
+        use supporting_functions_mod, only: energy_tot_func
 !
         implicit none
 !
@@ -919,7 +922,7 @@ endif
         !Save current z0 and stepped back global integration counter
         z_start_adaptive = intermediate_z0_list(:,number_of_integration_steps)
[…]
