+++ SRC/gorilla_plot_mod.f90
@@ -27,8 +27,7 @@ module gorilla_plot_mod
     & filename_orbit_start_pos_rphiz, filename_orbit_start_pos_sthetaphi, start_pos_x2, start_pos_x3, &
     & filename_e_tot, filename_p_phi, filename_J_par, start_pitch_parameter,energy_eV_start
 !
-    public              :: phi_elec_func, &
-                        & gorilla_plot
+    public              :: gorilla_plot
 !    
     contains
 !
@@ -419,6 +418,7 @@ module gorilla_plot_mod
                                                  & file_id_p_phi, file_id_J_par, file_id_full_orbit, counter_phi_0_mappings, &
                                                  & counter_vpar_0_mappings, counter_tetrahedron_passes)
 !
+            use supporting_functions_mod, only: bmod_func, energy_tot_func, vperp_func, p_phi_func
             use pusher_tetra_rk_mod, only: find_tetra,pusher_tetra_rk,initialize_const_motion_rk
             use pusher_tetra_poly_mod, only: pusher_tetra_poly,initialize_const_motion_poly
             use tetra_physics_poly_precomp_mod , only: make_precomp_poly_perpinv, initialize_boole_precomp_poly_perpinv, &
@@ -553,7 +553,7 @@ module gorilla_plot_mod
                         if(boole_e_tot) then
                             !Write elapsed time and total energy
                             !$omp critical
-                                write(file_id_e_tot,*) t_step - t_remain, E_tot_func(z_save,vpar,perpinv,ind_tetr_save)
+                                write(file_id_e_tot,*) t_step - t_remain, energy_tot_func([z_save,vpar],perpinv,ind_tetr_save)
                             !$omp end critical
                         endif !boole_e_tot
                     endif
@@ -613,7 +613,7 @@ module gorilla_plot_mod
 !
                             !Write number of toroidal mappings and total energy
                             !$omp critical
-                                write(file_id_e_tot,*) counter_phi_0_mappings, E_tot_func(z_save,vpar,perpinv,ind_tetr_save)
+                                write(file_id_e_tot,*) counter_phi_0_mappings, energy_tot_func([z_save,vpar],perpinv,ind_tetr_save)
                             !$omp end critical
                         endif
                     endif !boole_e_tot
@@ -639,126 +639,6 @@ module gorilla_plot_mod
 !             call alloc_precomp_poly_perpinv(2,ntetr)
 !         
         end subroutine gorilla_plot_orbit_integration
-!
-!ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
-!
-        function bmod_func(z123,ind_tetr)
-!
-            use tetra_physics_mod, only: tetra_physics
-!
-            implicit none
-!
-            double precision :: bmod_func
-            integer, intent(in) :: ind_tetr
-            double precision, dimension(3),intent(in) :: z123
-!
-            bmod_func = tetra_physics(ind_tetr)%bmod1+sum(tetra_physics(ind_tetr)%gb*z123)
-!        
-        end function bmod_func
-!
-!ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
-!
-        function vperp_func(z123,perpinv,ind_tetr)
-!        
-            use tetra_physics_mod, only: tetra_physics
-!    
-            implicit none
-!
-            double precision :: vperp_func
-            integer, intent(in) :: ind_tetr
-            double precision, dimension(3),intent(in) :: z123
-            double precision, intent(in) :: perpinv
-
[…]

[…]
