### 2023-02-16 bd2deec
Introduced v_E to tetraphysics quantities
