+++ SRC/gorilla_diag_mod.f90
@@ -6,7 +6,7 @@
     logical :: diag_pusher_tetra_rk = .false.
     logical :: transp_coef_trigger = .false.
     logical :: diag_mesh_SOLEDGE3X_EIRENE = .false.
-    logical :: diag_strong_electric_field = .true.
+    logical :: diag_strong_electric_field = .false.
   end module gorilla_diag_mod
 !
 !ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
