+++ SRC/tetra_physics_mod.f90
@@ -22,6 +22,10 @@
       double precision :: Phi1                   ! electrostatic potential in the first vertex
       double precision :: R1                     ! major radius in the first vertex
       double precision :: Z1                     ! Z in the first vertex
+      double precision :: vE1_1                  ! 1st (covariant) component of drift velcoity v_E in the first vertex
+      double precision :: vE2_1                  ! 2nd (covariant) component of drift velcoity v_E in the first vertex
+      double precision :: vE3_1                  ! 3rd (covariant) component of drift velcoity v_E in the first vertex
+      double precision :: v2Emod_1               ! module squared of drift velocity v_E in first vertex
       double precision :: Er_mod                 ! module of the electric field in r-direction
       double precision :: sqg1                   ! square root g (metric determinant) at the first vertex (ONLY for grid_kind = 3)
       double precision :: dt_dtau_const          ! Factor dt_dtau averaged of the four vertices
@@ -48,6 +52,10 @@
       double precision, dimension(3)   :: gh2    ! gradient of the 2nd component of the unit vector h
       double precision, dimension(3)   :: gh3    ! gradient of the 3rd component of the unit vector h
       double precision, dimension(3)   :: curlh  ! $\epsilon^{ijk}\difp{h_k}{x^j}$ = "curl" of $\bh$
+      double precision, dimension(3)   :: gvE1   ! gradient of the 1st component of the drift velocity v_E
+      double precision, dimension(3)   :: gvE2   ! gradient of the 2nd component of the drift velocity v_E
+      double precision, dimension(3)   :: gvE3   ! gradient of the 3rd component of the drift velocity v_E
+      double precision, dimension(3)   :: gv2Emod! gradient of the modulo squared of drift velocity v_E
       !3DGeoInt properties
       double precision, dimension(3,3) :: alpmat ! real space part of matrix alpha (block from 1 to 3)
       double precision, dimension(3,3) :: betmat ! real space part of matrix beta (block from 1 to 3)
@@ -107,7 +115,7 @@
       use gorilla_settings_mod, only: eps_Phi,handover_processing_kind, boole_axi_noise_vector_pot, &
             & boole_axi_noise_elec_pot, boole_non_axi_noise_vector_pot, axi_noise_eps_A, axi_noise_eps_Phi, &
             & non_axi_noise_eps_A, boole_strong_electric_fields
-      use strong_electric_field_mod, only: get_electric_field, save_electric_field
+      use strong_electric_field_mod, only: get_electric_field, save_electric_field, get_v_E
       use gorilla_diag_mod, only: diag_strong_electric_field
 !
       integer, intent(in) :: ipert_in,coord_system_in
@@ -122,7 +130,7 @@
       double precision, dimension(:), allocatable   :: davec_dx1,davec_dx2,davec_dx3
       double precision, dimension(:,:),allocatable :: avec
[…]
