+++ SRC/strong_electric_field_mod.f90
@@ -129,13 +129,15 @@ contains
         close(123)
 !
         !Stop program after successfully saving the E-field
-        STOP
+        !STOP
 !
     end subroutine save_electric_field
 !
 !ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
 !
     subroutine get_v_E(R,E_x1,E_x2,E_x3,h_x1,h_x2,h_x3,Bmod,v_E_x1,v_E_x2,v_E_x3,v2_E_mod)
+!
+        use constants, only: clight
 !
         implicit none
 !
@@ -145,9 +147,9 @@ contains
         !Writting out the cross-product for v_E (covariant components) using the cylindrical metric determinant g_lk = (1,R^2,1)
         !(v_E)_l = g_lk * epsilon^(kij) * E_i * h_j / (sqrt(g) * Bmod)
         !The factor c (speed of light) is left out for having unitless values (include later when determing energy and pusher)
-        v_E_x1 = (E_x2 * h_x3 - E_x3 * h_x2) / (R * Bmod)
-        v_E_x2 = (E_x3 * h_x1 - E_x1 * h_x3) / (Bmod) * R
-        v_E_x3 = (E_x1 * h_x2 - E_x2 * h_x1) / (R * Bmod)
+        v_E_x1 = (E_x2 * h_x3 - E_x3 * h_x2) / (R * Bmod) * clight
+        v_E_x2 = (E_x3 * h_x1 - E_x1 * h_x3) / (Bmod) * R * clight
+        v_E_x3 = (E_x1 * h_x2 - E_x2 * h_x1) / (R * Bmod) * clight
 !
         !Writing out the scalar-product of v_E with itself using the cylindrical metric determinant
         !v2_E_mod = (v_E)_l * (v_E)^l = (v_E)_l * g^{lk} * (v_E)_k
