+++ SRC/tetra_physics_mod.f90
@@ -33,14 +33,26 @@
                                                  ! of "curl A" with gradient of B module
       double precision :: gPhixcurlA             ! $\epsilon^{ijk}\difp{Phi}{x^i}\difp{A_k}{x^j}$ scalar
                                                  ! of "curl A" with gradient of Phi
+      double precision :: gv2EmodxcurlA          ! $\epsilon^{ijk}\difp{v2Emod}{x^i}\difp{A_k}{x^j}$ scalar
+                                                 ! of "curl A" with gradient of v2Emod
+      double precision :: gBxcurlvE              ! $\epsilon^{ijk}\difp{B}{x^i}\difp{vE_k}{x^j}$ scalar
+                                                 ! of "curl vE" with gradient of B module
+      double precision :: gPhixcurlvE            ! $\epsilon^{ijk}\difp{Phi}{x^i}\difp{vE_k}{x^j}$ scalar
+                                                 ! of "curl vE" with gradient of Phi
+      double precision :: gv2EmodxcurlvE          ! $\epsilon^{ijk}\difp{v2Emod}{x^i}\difp{vE_k}{x^j}$ scalar
+                                                 ! of "curl vE" with gradient of v2Emod
       double precision :: spalpmat               ! alpha matrix element(4,4)= trace of the real space
                                                  ! part of matrix alpha (see below)
       double precision :: spbetmat               ! beta matrix element(4,4)= trace of the real space
                                                  ! part of matrix beta (see below)
+      double precision :: spgammat               ! gamma matrix element(4,4)= trace of the real space
+                                                 ! part of matrix gamma (see below)
       double precision, dimension(3)   :: gBxh1  ! $\epsilon^{ijk} \difp{B}{x^j} h_k$ - vector product
                                                  ! of gradient B module with $\bh$ in the first vertex
       double precision, dimension(3)   :: gPhixh1  ! $\epsilon^{ijk} \difp{Phi}{x^j} h_k$ - vector product
                                                  ! of gradient Phi with $\bh$ in the first vertex
+      double precision, dimension(3)   :: gv2Emodxh1  ! $\epsilon^{ijk} \difp{v2Emod}{x^j} h_k$ - "vector product"
+                                                 ! of gradient v2Emod with $\bh$ in the first vertex
       double precision, dimension(3)   :: gB     ! gradient of B module
       double precision, dimension(3)   :: gPhi   ! gradient of Phi
       double precision, dimension(3)   :: gR     ! gradient of R (major radius)
@@ -55,10 +67,12 @@
       double precision, dimension(3)   :: gvE1   ! gradient of the 1st component of the drift velocity v_E
       double precision, dimension(3)   :: gvE2   ! gradient of the 2nd component of the drift velocity v_E
       double precision, dimension(3)   :: gvE3   ! gradient of the 3rd component of the drift velocity v_E
+      double precision, dimension(3)   :: curlvE ! $\epsilon^{ijk}\difp{vE_k}{x^j}$ = "curl" of $\bvE$
[…]
