### 2023-02-25 333bebe
Corrected curlvE and made potential consistent
