+++ README.md
@@ -55,9 +55,10 @@ To install requirements on Ubuntu Linux use
 sudo apt-get install wget unzip gfortran liblapack-dev libnetcdff-dev
 ```
 
-To install requirements on macOS, install [macports](https://www.macports.org/install.php), then
+To install requirements on macOS, install [Homebrew]([https://brew.sh](https://brew.sh)), then
 ```
-sudo port install netcdf netcdf-fortran
+brew install git gcc make cmake
+brew install netcdf netcdf-fortran
 ```
 ### Additional tools
 * [pFUnit](https://github.com/Goddard-Fortran-Ecosystem/pFUnit)
@@ -71,7 +72,7 @@ sudo apt install lcov
 ```
 To install lcov on macOS use
 ```
-sudo port install lcov
+brew install lcov
 ```
 
 ### Include external library
