+++ README.md
@@ -52,12 +52,12 @@ GORILLA can be built with `make` or `cmake`.
 
 To install requirements on Ubuntu Linux use
 ```
-sudo apt-get install wget unzip gfortran liblapack-dev libnetcdff-dev
+sudo apt install wget unzip git gfortran make cmake liblapack-dev libnetcdff-dev
 ```
 
 To install requirements on macOS, install [Homebrew]([https://brew.sh](https://brew.sh)), then
 ```
-brew install git gcc make cmake netcdf netcdf-fortran
+brew install wget unzip git gcc make cmake netcdf netcdf-fortran
 ```
 ### Additional tools
 * [pFUnit](https://github.com/Goddard-Fortran-Ecosystem/pFUnit)
