+++ SRC/pusher_tetra_poly.f90
@@ -110,7 +110,7 @@ endif
         subroutine initialize_pusher_tetra_poly(ind_tetr_inout,x,iface,vpar,t_remain_in)
 !
             use tetra_physics_mod, only: tetra_physics, sign_sqg
-            use gorilla_settings_mod, only: boole_strong_electric_fields
+            use gorilla_settings_mod, only: boole_strong_electric_field
             use supporting_functions_mod, only: bmod_func, phi_elec_func, v2_E_mod_func
 !
             implicit none
@@ -155,7 +155,7 @@ endif
 !
             k1 = vperp2+vpar2+2.d0*perpinv*tetra_physics(ind_tetr)%bmod1
             !Adding in +dx*grad(v2emod) to k1 helper-coefficiant for strong electric field case
-            if (boole_strong_electric_fields) k1 = k1 + (v2_E_mod_func(z_init(1:3),ind_tetr) - tetra_physics(ind_tetr)%v2Emod_1)
+            if (boole_strong_electric_field) k1 = k1 + (v2_E_mod_func(z_init(1:3),ind_tetr) - tetra_physics(ind_tetr)%v2Emod_1)
             k3 = tetra_physics(ind_tetr)%Phi1-phi_elec
 !
         end subroutine initialize_pusher_tetra_poly
@@ -1463,7 +1463,7 @@ if(diag_pusher_tetry_poly) print *, 'boole',boole_approx,'dtau',dtau,'iface_new'
         subroutine analytic_coeff_without_precomp(poly_order,boole_faces,z,coef_mat)
 !
             use tetra_physics_mod, only: tetra_physics,cm_over_e
-            use gorilla_settings_mod, only: boole_strong_electric_fields
+            use gorilla_settings_mod, only: boole_strong_electric_field
             use constants, only: clight
             use poly_without_precomp_mod
 !
@@ -1493,7 +1493,7 @@ if(diag_pusher_tetry_poly) print *, 'boole',boole_approx,'dtau',dtau,'iface_new'
                         - clight* tetra_physics(ind_tetr)%spbetmat
             amat(1:3,4) = tetra_physics(ind_tetr)%curlA
 !
-            if (boole_strong_electric_fields) then
+            if (boole_strong_electric_field) then
                 b(1:3) = b(1:3) - cm_over_e*tetra_physics(ind_tetr)%gv2Emodxh1 !and also the modification of k1 done by initialization
                 b(4) = b(4) + cm_over_e*perpinv*tetra_physics(ind_tetr)%gBxcurlvE - clight*tetra_physics(ind_tetr)%gPhixcurlvE &
                           & - 0.5d0*cm_over_e*tetra_physics(ind_tetr)%gv2EmodxcurlvE - 0.5d0*tetra_physics(ind_tetr)%gv2EmodxcurlA
@@ -2778,7 +2778,7 @@ if(diag_pusher_tetry_poly) print *, 'New quadratic solver is called.'
         use tetra_physics_poly_precomp_mod, only: tetra_physics_poly4,tetra_physics_poly1
         use tetra_physics_mod, only: tetra_physics, cm_over_e
         use poly_without_precomp_mod, only: b
-        use gorilla_settings_mod, only: i_precomp, boole_strong_electric_fields
+        use gorilla_settings_mod, only: i_precomp, boole_strong_electric_field
         use constants, only: clight
 !
         implicit none
@@ -2792,7 +2792,7 @@ if(diag_pusher_tetry_poly) print *, 'New quadratic solver is called.'
                         & perpinv*cm_over_e*tetra_physics_poly1(ind_tetr)%anorm_in_alpmat(:,iface)) * z(1:3)) + &
[…]

[…]
