### 2023-02-28 53c55e7
Added v_E diagnostic and fixed formula error
