### 2023-03-01 0a6997b
Added control options and revised ExB for tau_est
