### 2023-03-02 b272298
Implemented strong electric fields for RK4
