+++ SRC/pusher_tetra_rk.f90
@@ -53,8 +53,9 @@ module pusher_tetra_rk_mod
         use tetra_grid_settings_mod, only: grid_size
         use supporting_functions_mod, only: bmod_func
         use constants, only : clight,eps,pi
-        use gorilla_settings_mod, only: boole_dt_dtau
+        use gorilla_settings_mod, only: boole_dt_dtau, boole_strong_electric_field
         use gorilla_diag_mod, only: diag_pusher_tetra_rk
+        use supporting_functions_mod, only: v2_E_mod_func, phi_elec_func
 !
         implicit none
 !
@@ -93,7 +94,7 @@ module pusher_tetra_rk_mod
         bmod=bmod_func(z_init(1:3),ind_tetr)
 !
         !Phi at the entry point of the particle
-        phi_elec=tetra_physics(ind_tetr)%Phi1+sum(tetra_physics(ind_tetr)%gPhi*z_init(1:3))
+        phi_elec = phi_elec_func(z_init(1:3),ind_tetr)
 !
         !Auxiliary quantities
         vperp2 = -2.d0*perpinv*bmod
@@ -120,6 +121,16 @@ module pusher_tetra_rk_mod
         spamat=perpinv*cm_over_e*tetra_physics(ind_tetr)%spalpmat &
             - clight* tetra_physics(ind_tetr)%spbetmat    !tr(a-Matrix)
 !
+        if (boole_strong_electric_field) then
+            b(1:3) = b(1:3) - 0.5d0*cm_over_e*tetra_physics(ind_tetr)%gv2Emodxh1 & !In RK4 the k1 coefficiant is written out explicitly
+                 & +cm_over_e*tetra_physics(ind_tetr)%curlh * (v2_E_mod_func(z_init(1:3),ind_tetr)-tetra_physics(ind_tetr)%v2Emod_1)
+            b(4) = b(4) + cm_over_e*perpinv*tetra_physics(ind_tetr)%gBxcurlvE - clight*tetra_physics(ind_tetr)%gPhixcurlvE &
+                        & - 0.5d0*cm_over_e*tetra_physics(ind_tetr)%gv2EmodxcurlvE - 0.5d0*tetra_physics(ind_tetr)%gv2EmodxcurlA
+            amat = amat - 0.5d0*cm_over_e*tetra_physics(ind_tetr)%gammat
+            spamat = spamat - 0.5d0*cm_over_e*tetra_physics(ind_tetr)%spgammat !amat(4,4) in pusher_tetra_poly case
+            Bvec = Bvec + cm_over_e*tetra_physics(ind_tetr)%curlvE !amat(1:3,4) in pusher_tetra_poly_case
+        endif !boole_stron_electric_fields (adding additional terms to the coefficiants)
+!
         !Multiply A-matrix (4x4) and b with appropriate sign (which ensures that tau remains positive inside the algorithm)
         amat = amat * dble(sign_rhs)
         b = b * dble(sign_rhs)
@@ -135,7 +146,12 @@ module pusher_tetra_rk_mod
         tetra_dist_ref = abs(tetra_physics(ind_tetr)%tetra_dist_ref)
 !        
         !ExB drift velocity
-        vd_ExB = abs(clight/bmod*tetra_physics(ind_tetr)%Er_mod)
+        !Only in case of strong electric field mode, ExB drift directly calculated during tetra_physics-setup
+        if (boole_strong_electric_field) then
+            vd_ExB = tetra_physics(ind_tetr)%v_E_mod_average
+        else
+            vd_ExB = abs(clight/bmod*tetra_physics(ind_tetr)%Er_mod)
+        endif
         boole_vd_ExB = .true.
[…]
