### 2023-03-04 48cde3a
Considered MOVING FRAME for vmod calculation
