### 2023-03-04 e780118
Started work on demonstration example for E-fields
