+++ MATLAB/example_8.m
@@ -1,0 +1,603 @@
+%#######################################################################################################################
+% description:
+%-----------------------------------------------------------------------------------------------------------------------
+% * Using a potential that is a scaled poloidal flux (ASDEX-mesh) calculating the corresponding electric field via central differences.
+% * Perform each a run with and without the inclusion of the additional terms in the strong electric field Lagrangian using fully ionised Tungsten.
+% * Plot the poincare cross-section of both runs, as well as their respective fluctuations of total energy and toroidal momentum.
+% * The definition of total energy and toroidal momentum is different in the two runs. Each definition should yield conserved quantities in their respective case.
+%
+%#######################################################################################################################
+% used functions:
+%-----------------------------------------------------------------------------------------------------------------------
+% InputFile
+% NameList
+% read_in
+% load_copy
+%#######################################################################################################################
+% authors: Georg Graßler
+% created: 04.03.2023
+
+
+%Initialize used paths and files
+%-----------------------------------------------------------------------------------------------------------------------
+
+%Name of the current calculation to create folder
+name_test_case='example_8';
+
+%Control quantitites
+%close all
+
+lambda(1) = -0.4; %0.3 for WEST / -0.9 for ASDEX
+eps_Phi = 2*1e-7; %-4*1e-4 for WEST / +1e-4 for ASDEX / +1e-5 ASDEX Tungsten
+%eps_Phi = 1e-7;
+energy_eV_start = 3e3; %3e3
+time_multiplier = 100;
+boole_strong_electric_field = true;
+boole_calculate_conserved_quantities = false;
+chosen_orbit_index =10; %8/11 for strong_electric else 8/10
+grid_kind = 2;
+ipusher = 2;
+ispecies = 4;
+start_ASDEX = [210,0];
+
+standart = 1e-7;
+standart_energy = 3e3;
+standart_lambda = 0.3;
+%total_orbit_time = 0.002*standart/max(abs(eps_Phi),standart)*sqrt(standart_energy/energy_eV_start);
+total_orbit_time = 2e-5;
+
+if (boole_calculate_conserved_quantities)
+    total_orbit_time = time_multiplier*total_orbit_time;
+    n_skip_full_orbit = time_multiplier;
+    i_orbit_options = 1;
+else
+    n_skip_full_orbit = 1;
+    i_orbit_options = 3;
+end
+
+if (grid_kind == 2)
+    energy_eV_start = 3e5;
+    g_file_filename = 'MHD_EQUILIBRIA/g_file_for_test';
+    convex_wall_filename = 'MHD_EQUILIBRIA/convex_wall_for_test.dat';
+else
+    g_file_filename = 'MHD_EQUILIBRIA/g_file_for_test_WEST';
+    convex_wall_filename = 'MHD_EQUILIBRIA/convex_wall_for_test_WEST.dat';
+end
+
+%path of Matlab script
+path_script=pwd;
+
+%main path of GORILLA
+c=strsplit(path_script,'/');
+path_main=strjoin(c(1:end-1),'/');
+
+%path to run GORILLA code
+mkdir(path_main,['EXAMPLES/MATLAB_RUN/',name_test_case]);
+path_RUN=[path_main,'/EXAMPLES/MATLAB_RUN/',name_test_case];
+
[…]
