+++ .github/workflows/Mac.yml
@@ -125,6 +125,7 @@ jobs:
             example_5
             example_6
             example_7
+            example_8
 
       - name: Run MATLAB livescript
         uses: matlab-actions/run-command@v1
