+++ MATLAB/example_8.m
@@ -26,11 +26,6 @@ name_test_case='example_8';
 
 %Control quantitites
 close all
-
-%lambda(1) = -0.9; %0.3 for WEST / -0.9 for ASDEX -0.4
-%eps_Phi = 2*1e-5; %-4*1e-4 for WEST / +1e-4 for ASDEX / +1e-5 ASDEX Tungsten
-%eps_Phi = 1e-7;
-
 grid_kind = 2;
 
 %Setting configurations depending on test case
@@ -45,11 +40,11 @@ switch(grid_kind)
         g_file_filename = 'MHD_EQUILIBRIA/g_file_for_test';
         convex_wall_filename = 'MHD_EQUILIBRIA/convex_wall_for_test.dat';
     case(4)
-        eps_Phi = -4*1e-4;
-        starting_2D = [210,38];
-        lambda = 0.3;
+        eps_Phi = -1.5*1e-5;
+        starting_2D = [287,2];
+        lambda = -0.9;
         n2 = 10;
-        energy_eV_start = 3e5;
+        energy_eV_start = 6e5;
         total_orbit_time = 2e-2;
         g_file_filename = 'MHD_EQUILIBRIA/g_file_for_test_WEST';
         convex_wall_filename = 'MHD_EQUILIBRIA/convex_wall_for_test_WEST.dat';
@@ -153,11 +148,11 @@ tetra_grid.read();
 
     %Grid Size
         %Rectangular: nR, Field-aligned: ns
-        tetra_grid.TETRA_GRID_NML.n1 = 40;
+        tetra_grid.TETRA_GRID_NML.n1 = 20;
         %Rectangular: nphi, Field-aligned: nphi
         tetra_grid.TETRA_GRID_NML.n2 = n2;
         %Rectangular: nZ, Field-aligned: ntheta
-        tetra_grid.TETRA_GRID_NML.n3 = 40;
+        tetra_grid.TETRA_GRID_NML.n3 = 60;
 
     %Grid kind
         %1 ... rectangular grid for axisymmetric EFIT data
@@ -298,8 +293,8 @@ cd(path_RUN);
 
 %Color scheme and plotting settings
 grid_thickness = 0.2;
-electric_thickness = 0.2;
-drift_thickness = 0.2;
+electric_thickness = 1;
+drift_thickness = 1;
 
 grid_color = [204,204,204]/256;
 electric_field_color = [256,0,0]/256;
@@ -450,20 +445,10 @@ for t = [1:n_faces]
     v_E_tor_grid((t-1)*5 + 5,:) = nan;
 end
 
-%Calculate Mach number
-clight = 3*1e10;
-mass_tungsten = 184 * 1.675 * 1e-24;
-eV2erg = 1.60218*1e-12;
-
-v_E_mod_average = sum(v_E_mod,1)/size(v_E_mod,1);
-v_thermal = sqrt(2*energy_eV_start*eV2erg/mass_tungsten);
-mach_number_average = v_E_mod_average/v_thermal;
-mach_number_max = max(v_E_mod)/v_thermal;
-
 
 %%Create Plots
 %-----------------------------------------------------------------------------------------------------------------------
-close all
+
 %Draw Poincare plot
 xlabel_txt = '$R$ [cm]';
 ylabel_txt = '$Z$ [cm]';
@@ -633,56 +618,3 @@ function [v_cart] = cylinder2cart(v_cylinder,r,phi)
     v_cart(:,2) = v_r.*sin(phi) + v_phi./r.*cos(phi);
     v_cart(:,3) = v_z;
 end
-
-%% REST
-%     grid_plot = plot3(grid(:,1),zeros(size(grid(:,1))),grid(:,2),'Color',grid_color);
-%     grid_plot.LineWidth = grid_thickness;
[…]
