+++ README.md
@@ -154,7 +154,7 @@ However, GORILLA itself can be run without these packages. They are only used to
 
 ## Examples
 
-Five examples for plotting Poincaré cuts, full guiding-center orbits (in plasma core or edge regions) and the appropriate time evolution of invariants of motion can be found in `EXAMPLES/example_1` - `EXAMPLES/example_7`. There, the necessary soft links are already created and the input files are given, and runs are started with
+Seven examples for plotting Poincaré cuts, full guiding-center orbits (in plasma core or edge regions) and the appropriate time evolution of invariants of motion can be found in `EXAMPLES/example_1` - `EXAMPLES/example_7`. There, the necessary soft links are already created and the input files are given, and runs are started with
 ```
 ./test_gorilla_main.x   #if the build was done with make
 ```
@@ -165,8 +165,8 @@ or
 To avoid hyperthreading issues, it is beneficial to limit the number of threads to
 the number of actual CPU cores via the environment variable `$OMP_NUM_THREADS`.
 Detailed descriptions of the respective input files can be found in `INPUT`.
-After appropriate compilation of GORILLA, the code can be executed in all of these 6 example folders, respectively.
-For the visualization of the output of these five examples, appropriate plotting methods for Python 3 are at disposal at `PYTHON/plot_example_1.py` - `PYTHON/plot_example_7.py`.
+After appropriate compilation of GORILLA, the code can be executed in all of these 7 example folders, respectively.
+For the visualization of the output of these seven examples, appropriate plotting methods for Python 3 are at disposal at `PYTHON/plot_example_1.py` - `PYTHON/plot_example_7.py`.
 
 ### Example 1
 * Compute a collisionless guiding-center orbit with GORILLA for a trapped Deuterium particle.
@@ -211,10 +211,16 @@ For the visualization of the output of these five examples, appropriate plotting
  * Compute the normalized parallel adiabatic invariant as a function of banana bounces.
  * Plot fluctuation and evolution of energy over the bounces.
 
+ ### Example 8: (Currently only available via the MATLAB script example_8.m)
+ * Using a potential that is a scaled poloidal flux (ASDEX/SOLEDGE3X-mesh), calculate the corresponding electric field via central differences.
+ * Perform two runs, one with and one without the inclusion of the additional terms in the strong electric field Lagrangian using ionised Tungsten as particle species.
+ * Plot the poincare cross-section of both runs, their respective fluctuations of total energy and toroidal momentum. as well as the electric field and ExB-drift.
+ * The definition of total energy and toroidal momentum is different in the two runs. Each definition should yield conserved quantities in their respective case.
+
 
 ### Generation of input files and plotting in MATLAB and Python
 A detailed explanation of all examples (1-7) including the generation of the appropriate input files (including the example folders in `EXAMPLES/MATLAB_RUN` and `EXAMPLES/PYTHON_RUN`) and plotting of the results with MATLAB and Python can be found in the folders `MATLAB` and `PYTHON`, respectively.
-Here, the results of GORILLA with different polynominal orders K=2,3,4 and Runge-Kutta 4 are compared in case of examples 1-3. For examples 5-6 orbits for both trapped and passing particles are calculated. For example 7 an additional, in-depth comparison between adaptive and non-adaptive scheme is performed.
+Here, the results of GORILLA with different polynominal orders K=2,3,4 and Runge-Kutta 4 are compared in case of examples 1-3. For examples 5-6 orbits for both trapped and passing particles are calculated. For example 7 an additional, in-depth comparison between adaptive and non-adaptive scheme is performed. The last example, example 8, is currently only available via the MATLAB script `example_8.m`. A redundant version in PYTHON, as well as a corresponding example-folder with an appropriate script for plotting will follow in the future.
 
 
 ## Tests and coverage
