+++ SRC/pusher_tetra_rk_before_adding_grid_for_find_tetra.f90
@@ -1,3024 +1,0 @@
-!
-module pusher_tetra_rk_mod
-!
-    private
-!
-    integer                          :: iface_init
-    integer,public,protected         :: ind_tetr
-    double precision                 :: B0,perpinv,perpinv2,vmod_init
-    double precision,public,protected :: spamat,dt_dtau_const
-    double precision                 :: dist1,dist_min,dist_max,t_remain
-    double precision, dimension(3)   :: gradB,x_init
-    double precision, dimension(3),public,protected   :: Bvec
-    double precision, dimension(4),public,protected   :: b,z_init
-    double precision, dimension(3,3),public,protected :: amat
-    double precision, dimension(3,4) :: anorm
-    double precision,parameter                 :: eps_distmin = 1.d-10
-    double precision,parameter                 :: eps_distmax = 10.d0
-    double precision                 :: dtau_ref,dtau_max,dtau_quad
-    integer, parameter               :: kiter=48  
-    double precision,parameter      :: eps_modulation = 0.1d0
-    double precision,parameter      :: eps_dtau_max = 10.d0
-    double precision,parameter      :: eps_dtau_quad = 1.5d0
-    double precision                 :: k1, k3
-!   
-    !$OMP THREADPRIVATE(ind_tetr,spamat,B0,perpinv,perpinv2,iface_init,dt_dtau_const,k1,k3, &
-    !$OMP& dist1,dist_min,dist_max,Bvec,gradB,x_init,b,z_init,amat,anorm,t_remain,vmod_init,dtau_ref,dtau_max,dtau_quad)
-!    
-    public :: pusher_tetra_rk,find_tetra,initialize_const_motion_rk,energy_tot_func
-!    
-    contains
-!
-!ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
-!
-    subroutine initialize_const_motion_rk(perpinv_in,perpinv2_in)
-!
-        implicit none
-!            
-        double precision, intent(in)    :: perpinv_in,perpinv2_in
-!
-        perpinv = perpinv_in
-        perpinv2 = perpinv2_in
-!            
-    end subroutine initialize_const_motion_rk
-!
-!ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
-!
-    subroutine initialize_pusher_tetra_rk_mod(ind_tetr_inout,x,iface,vpar,t_remain_in)
-!
-        use tetra_physics_mod, only: tetra_physics, cm_over_e,dt_dtau,coord_system
-        use tetra_grid_settings_mod, only: grid_size
-        use constants, only : clight,eps,pi
-        use gorilla_settings_mod, only: boole_dt_dtau
-        use gorilla_diag_mod, only: diag_pusher_tetra_rk
-!
-        implicit none
-!
-        integer, intent(in)                        :: ind_tetr_inout,iface
-        double precision, intent(in)               :: vpar
-        double precision, dimension(3), intent(in) :: x
-        double precision                           :: bmod,vperp2,vpar2,phi_elec,t_remain_in,tetra_dist_ref,vd_ExB
-        logical                                    :: boole_vd_ExB,boole_vperp
-!
-        t_remain = t_remain_in
-!    
-        ind_tetr=ind_tetr_inout           !Save the index of the tetrahedron locally
-!
-        z_init(1:3)=x-tetra_physics(ind_tetr)%x1       !x is the entry point of the particle into the tetrahedron in (R,phi,z)-coordinates
-!
-        z_init(4)=vpar                         !Transform to z_init: 1st vertex of tetrahdron is new origin
-!
-        !Save initial orbit parameters
-        x_init = x
[…]
