+++ SRC/pusher_tetra_poly_version _without_interface.f90
@@ -1,2982 +1,0 @@
-!
-module poly_without_precomp_mod
-!
-    implicit none
-!
-    double precision, dimension(4)      :: b
-    double precision, dimension(4,4)    :: amat,amat2,amat3,amat4
-    double precision, dimension(4)      :: amat_in_z,amat2_in_z,amat3_in_z,amat_in_b,amat2_in_b
-    double precision, dimension(4)      :: amat4_in_z,amat3_in_b
-    double precision                    :: perpinv3,perpinv4
-!    
-    !$OMP THREADPRIVATE(b,amat,amat2,amat3,amat4,amat_in_z,amat2_in_z,amat3_in_z,amat4_in_z, & 
-    !$OMP& amat_in_b,amat2_in_b,amat3_in_b,perpinv3,perpinv4)
-!    
-end module poly_without_precomp_mod
-!
-!ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
-!
-module pusher_tetra_poly_mod
-
-    implicit none
-!
-    private
-!    
-    integer                             :: iface_init
-    integer, public, protected          :: ind_tetr
-    double precision                    :: perpinv2,vmod0
-    double precision, public, protected :: perpinv,dt_dtau_const,bmod0
-    double precision                    :: t_remain
-    double precision, dimension(3)      :: x_init
-    double precision, dimension(4), public, protected :: z_init
-    double precision                    :: k1, k3
-    double precision,parameter          :: eps_tau = 100.d0
-    double precision, dimension(4,4)    :: unity_matrix4 = reshape( [ 1.d0, 0.d0, 0.d0, 0.d0, &
-                               0.d0, 1.d0, 0.d0, 0.d0, & 
-                               0.d0, 0.d0, 1.d0, 0.d0, &
-                               0.d0, 0.d0, 0.d0, 1.d0 ], [4,4])
-!
-    !change those for adaptive step sizes, probably allocatable
-    double precision, dimension(:), allocatable           :: tau_steps_list
-    double precision, dimension(:,:), allocatable         :: intermediate_z0_list
-    integer                                               :: number_of_integration_steps
-!
-    !Diagnostics for adaptive step scheme (only useable for one particle calculation)
-    double precision, dimension(:,:), allocatable         :: total_fluctuation_report, single_step_fluctuation_report, &
-                                                           & closure_fluctuation_report
-    integer                                               :: report_entry_index, number_reports
-    integer, parameter                                    :: max_number_reports = 1, minimum_partition_number = 9000
-    logical                                               :: boole_collect_data
-!
-    !$OMP THREADPRIVATE(ind_tetr,iface_init,perpinv,perpinv2,dt_dtau_const,bmod0,t_remain,x_init,  &
-    !$OMP& z_init,k1,k3,vmod0,tau_steps_list,intermediate_z0_list,number_of_integration_steps)
-!
-    public :: pusher_tetra_poly,initialize_const_motion_poly, &
-        & Quadratic_Solver2, Cubic_Solver, Quartic_Solver,analytic_integration_without_precomp,energy_tot_func
-!   
-    contains
-!
-!ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
-!
-        subroutine initialize_const_motion_poly(perpinv_in,perpinv2_in)
-!
-            implicit none
-!            
-            double precision, intent(in)    :: perpinv_in,perpinv2_in
-!
-            perpinv = perpinv_in
-            perpinv2 = perpinv2_in
-!            
-        end subroutine initialize_const_motion_poly
-!
-!ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
[…]
