+++ SRC/pusher_tetra_poly.f90
@@ -2162,6 +2162,7 @@ if(diag_pusher_tetry_poly) print *, 'New quadratic solver is called.'
         use gorilla_settings_mod, only: boole_time_Hamiltonian, boole_gyrophase, boole_vpar_int, boole_vpar2_int, &
                                         & optional_quantities_type
         use tetra_physics_mod, only: hamiltonian_time,cm_over_e,tetra_physics
+        use tetra_grid_mod, only: verts_rphiz, tetra_grid
 !
         implicit none
 !
@@ -2175,124 +2176,71 @@ if(diag_pusher_tetry_poly) print *, 'New quadratic solver is called.'
         double precision, dimension(3)                  :: x0
         double precision                                :: vpar0
         double precision, dimension(:,:), allocatable   :: x_coef,x_vpar_coef
-        double precision, dimension(:), allocatable     :: vpar_coef,res_poly_coef
+        double precision, dimension(:), allocatable     :: vpar_coef,res_poly_coef,r_rescaled_coef
         double precision, dimension(4,poly_order+1)     :: dummy_coef_mat !exists because of analytic_coeff_without_precomp syntax
         logical, dimension(4)                           :: boole_faces_off = .false. !avoid root coefficients computation
+        double precision, dimension(:), allocatable     :: dtdtau_coef, omega_coef
+        integer                                         :: i
 !
         !recalculate polynomial coefficients (tensors) if more than one integration step was performed
         if (number_of_integration_steps .gt. 1) call set_integration_coef_manually(poly_order,z0)
 !
-            allocate(x_coef(3,poly_order+1))
-            allocate(vpar_coef(poly_order+1))
+        allocate(x_coef(3,poly_order+1))
+        allocate(vpar_coef(poly_order+1))
+        allocate(dtdtau_coef(poly_order+1))
+        allocate(x_vpar_coef(3,poly_order+1))
 !
-            call z_series_coef(poly_order,z0,x_coef,vpar_coef)
+        call z_series_coef(poly_order,z0,x_coef,vpar_coef)
+        x_vpar_coef = poly_multiplication(vpar_coef,x_coef)
 !
-        if(boole_time_hamiltonian.or.boole_vpar_int.or.boole_vpar2_int) then
-!
-            allocate(x_vpar_coef(3,poly_order+1))
-!
-            call poly_multiplication_coef(x_coef(1,:),vpar_coef(:),x_vpar_coef(1,:))
-            call poly_multiplication_coef(x_coef(2,:),vpar_coef(:),x_vpar_coef(2,:))
-            call poly_multiplication_coef(x_coef(3,:),vpar_coef(:),x_vpar_coef(3,:))
-        endif
+        do i = 1,poly_order+1
+            dtdtau_coef(i) = sum(hamiltonian_time(ind_tetr)%vec_mismatch_der * x_coef(:,i)) + &
+                            & cm_over_e * hamiltonian_time(ind_tetr)%h1_in_curlh * vpar_coef(i) + &
+                            & cm_over_e * sum(hamiltonian_time(ind_tetr)%vec_parcurr_der * x_vpar_coef(:,i))
+        enddo
+        dtdtau_coef(1) = dtdtau_coef(1) + hamiltonian_time(ind_tetr)%h1_in_curlA        
 !
         !Optional computation of Hamiltonian time
         if(boole_time_hamiltonian) then
-            delta_t_hamiltonian = hamiltonian_time(ind_tetr)%h1_in_curlA * tau + &
-            & cm_over_e * hamiltonian_time(ind_tetr)%h1_in_curlh * moment_integration(poly_order,tau,vpar_coef)+&
[…]
