+++ .github/workflows/Mac.yml
@@ -13,9 +13,9 @@ jobs:
     runs-on: macos-latest
     
     env:
-      CC: /opt/homebrew/bin/gcc
-      CXX: /opt/homebrew/bin/gcc
-      FC: /opt/homebrew/bin/gfortran
+      CC: /usr/local/bin/gcc
+      CXX: /usr/local/bin/gcc
+      FC: /usr/local/bin/gfortran
       PFUNIT_DIR: /Users/runner/work/GORILLA/GORILLA/pFUnit/build/installed/PFUNIT-4.6
 
     steps:
@@ -60,9 +60,9 @@ jobs:
     runs-on: macos-latest
     
     env:
-      CC: gcc
-      CXX: gcc
-      FC: gfortran
+      CC: /usr/local/bin/gcc
+      CXX: /usr/local/bin/gcc
+      FC: /usr/local/bin/gfortran
       PFUNIT_DIR: /home/runner/work/GORILLA/GORILLA/pFUnit/build/installed/PFUNIT-4.6
 
     steps:
