+++ .github/workflows/Mac.yml
@@ -13,9 +13,9 @@ jobs:
     runs-on: macos-latest
     
     env:
-      CC: $HOMEBREW_PREFIX/bin/gcc
-      CXX: $HOMEBREW_PREFIX/bin/gcc
-      FC: $HOMEBREW_PREFIX/bin/gfortran
+      CC: /usr/local/opt/bin/gcc
+      CXX: /usr/local/opt/bin/gcc
+      FC: /usr/local/opt/bin/gfortran
       PFUNIT_DIR: /Users/runner/work/GORILLA/GORILLA/pFUnit/build/installed/PFUNIT-4.6
 
     steps:
@@ -60,9 +60,9 @@ jobs:
     runs-on: macos-latest
     
     env:
-      CC: $HOMEBREW_PREFIX/bin/gcc
-      CXX: $HOMEBREW_PREFIX/bin/gcc
-      FC: $HOMEBREW_PREFIX/bin/gfortran
+      CC: /usr/local/opt/bin/gcc
+      CXX: /usr/local/opt/bin/gcc
+      FC: /usr/local/opt/bin/gfortran
       PFUNIT_DIR: /home/runner/work/GORILLA/GORILLA/pFUnit/build/installed/PFUNIT-4.6
 
     steps:
