+++ .github/workflows/Mac-make.yml
@@ -27,6 +27,8 @@ jobs:
       - name: Dependencies
         run: |
           brew install wget unzip git gcc make cmake netcdf netcdf-fortran
+          echo $HOMEBREW_PREFIX
+          echo $PATH
           
       - name: Additional files
         run: |
