+++ .github/workflows/Mac-make.yml
@@ -27,7 +27,8 @@ jobs:
       - name: Dependencies
         run: |
           brew install wget unzip git gcc make cmake netcdf netcdf-fortran
-          echo $HOMEBREW_PREFIX
+          env
+          env | grep HOMEB
           echo $PATH
           
       - name: Additional files
