+++ .github/workflows/Mac-make.yml
@@ -27,9 +27,7 @@ jobs:
       - name: Dependencies
         run: |
           brew install wget unzip git gcc make cmake netcdf netcdf-fortran
-          env
-          env | grep HOMEB
-          echo $PATH
+          brew config
           
       - name: Additional files
         run: |
