+++ .github/workflows/Mac.yml
@@ -14,7 +14,7 @@ jobs:
     
     env:
       FC: /usr/local/bin/gfortran
-      PFUNIT_DIR: /Users/runner/work/GORILLA/GORILLA/pFUnit/build/installed/PFUNIT-4.6
+      PFUNIT_DIR: /Users/runner/work/GORILLA/GORILLA/pFUnit/build/installed/PFUNIT-4.7
 
     steps:
       - name: Checkout
@@ -59,7 +59,7 @@ jobs:
     
     env:
       FC: /usr/local/bin/gfortran
-      PFUNIT_DIR: /home/runner/work/GORILLA/GORILLA/pFUnit/build/installed/PFUNIT-4.6
+      PFUNIT_DIR: /home/runner/work/GORILLA/GORILLA/pFUnit/build/installed/PFUNIT-4.7
 
     steps:
       - name: Checkout
