+++ .github/workflows/Mac.yml
@@ -13,6 +13,7 @@ jobs:
     runs-on: macos-latest
     
     env:
+      CC: /usr/local/bin/gcc
       FC: /usr/local/bin/gfortran
       PFUNIT_DIR: /Users/runner/work/GORILLA/GORILLA/pFUnit/build/installed/PFUNIT-4.7
 
@@ -58,6 +59,7 @@ jobs:
     runs-on: macos-latest
     
     env:
+      CC: /usr/local/bin/gcc
       FC: /usr/local/bin/gfortran
       PFUNIT_DIR: /home/runner/work/GORILLA/GORILLA/pFUnit/build/installed/PFUNIT-4.7
 
