### 2023-05-13 a600781
Fixes JOSS (#14)
* Fixed CMake and Make for macos arm64

* Change docs to Homebrew

* Fix GitHub actions CI
