+++ .github/workflows/Mac-make.yml
@@ -16,7 +16,7 @@ jobs:
     runs-on: macos-latest
     
     env:
-      FC: gfortran-9
+      FC: /usr/local/bin/gfortran
 
     # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
@@ -26,9 +26,7 @@ jobs:
 
       - name: Dependencies
         run: |
-          curl -LO https://raw.githubusercontent.com/GiovanniBussi/macports-ci/master/macports-ci
-          source ./macports-ci install
-          sudo port install netcdf netcdf-fortran
+          brew reinstall wget unzip git gcc make cmake netcdf netcdf-fortran
           
       - name: Additional files
         run: |
@@ -36,17 +34,10 @@ jobs:
           wget -O 954.zip "https://dl.acm.org/action/downloadSupplement?doi=10.1145%2F2699468&file=954.zip&download=true"
           unzip 954.zip
           cp 954/F90/Src/Polynomial234RootSolvers.f90 SRC/contrib/
-          
-      - name: Provide gfortran9 
-        run: |
-          sudo ln -s /usr/local/bin/gfortran-9 /usr/local/bin/gfortran 
-          sudo mkdir /usr/local/gfortran
-          sudo ln -s /usr/local/Cellar/gcc@9/9.3.0_1/lib/gcc/9 /usr/local/gfortran/lib
-          gfortran --version
         
       - name: Build
         run: |
-          make
+          /usr/local/bin/gmake
           
       - name: Run examples
         run: |
