+++ README.md
@@ -1,5 +1,5 @@
 [![DOI](https://zenodo.org/badge/343767192.svg)](https://zenodo.org/badge/latestdoi/343767192)
-[![](https://img.shields.io/github/workflow/status/itpplasma/GORILLA/CI)](https://github.com/itpplasma/GORILLA/actions/workflows/Ubuntu.yml)
+[![](https://img.shields.io/github/actions/workflow/status/itpplasma/GORILLA/Ubuntu.yml?branch=main)](https://github.com/itpplasma/GORILLA/actions/workflows/Ubuntu.yml)
 [![GitHub](https://img.shields.io/github/license/itpplasma/GORILLA?style=flat)](https://github.com/itpplasma/GORILLA/blob/master/LICENCE)
 [![status](https://joss.theoj.org/papers/842b83a430e4c2985d202c79ed626587/status.svg)](https://joss.theoj.org/papers/842b83a430e4c2985d202c79ed626587)
 
