### 2023-05-17 #4890 @krystophny
@pibion @rogeriojorge we have now fixed the issues:

* Mac build is working (coverage is only generated for Ubuntu now)
* Comments have been added to tests
* Coverage badge was skipped for now but we will add it in the future

Do you think that this version is now OK to be published?
