### 2023-05-22 #4890 @pibion
@krystophny some comments on the paper:

## Author ORCIDs
Not all the authors have ORCID IDs - this isn't a requirement for publication, but now is a good time to review the list and see if any of those missing an ORCID ID do in fact have them

## General comments
- line 15, A future use of this energy source ... -> This high energy density has made fusion the subject of worldwide research projects
- line 20, utilizes -> uses
- line 42, utilize -> use
- line 43, maybe omit "special" or be more specific about what is special?
- line 50, Such exemplary -> Two examples of guiding-center orbits obtained with GORILLA ...

## Some of the statements in the "statment of need" section are not immediately clear to me
- My impression is that GORILLA can provide phase-space coordinates at a specified time step to either a user or other programs, and that it does so quickly. In addition, in contrast to the many programs that can trace particles in the core,  GORILLA can trace particles both in the core and in the scrape-off layer.  Is that correct?  Also, is it unique in being able to provide phase-space coordinates, or is the ease of the API unique?  Finally, is it unique in its ability to conserve invariants?

- line 56, specifically, the initial set of coordinates is provided -> specifically, the user provides the initial set of coordinates?
- line 58, the main interest is to retrieve -> GORILLA's main purpose is to retrieve
- I'm a little confused about the integration process.  On line 59 you say it's irrelevant, but on line 61 you talk about it being of interest and providing tools to explore its effects (I think).  Could you say more about this?
