### 2023-05-26 d234f2a
Small corrections JOSS
