+++ PAPER_JOSS/paper.bib
@@ -316,3 +316,17 @@
 	pages = {P3.1059},
 	file = {Eder et al. - Integration of the guiding-center equations in tor.pdf:/Users/micheder/Zotero/storage/NKCA46JJ/Eder et al. - Integration of the guiding-center equations in tor.pdf:application/pdf},
 }
+
+@article{kasilov_geometric_2016,
+title = {Geometric integrator for charged particle orbits in axisymmetric fusion devices},
+journal = {Computer Physics Communications},
+volume = {207},
+pages = {282-286},
+year = {2016},
+issn = {0010-4655},
+doi = {https://doi.org/10.1016/j.cpc.2016.07.019},
+url = {https://www.sciencedirect.com/science/article/pii/S0010465516302077},
+author = {S.V. Kasilov and A.M. Runov and W. Kernbichler},
+keywords = {Plasma physics, Kinetic modelling, Numerical integrators},
+abstract = {A semi-analytical geometric integrator of guiding centre orbits in an axisymmetric tokamak is described. The integrator preserves all three invariants of motion up to computer accuracy at the expense of reduced orbit accuracy and it is roughly an order of magnitude more efficient than a direct solution of the equations of guiding centre motion with a standard high order adaptive ODE integrator.}
+}
