### 2023-05-26 #4890 @krystophny
Thank you, @pibion !

We have adressed your comments in the latest version.

1) ORCID: Indeed only two of the authors have an ORCID, as the others are either students or already close to retirement.

2) Minor remarks in the text: We have corrected them

3) Statement of need: We have briefly commented on the unique feature of the approach as follows in the text: A unique feature of `GORILLA` is the direct way in which fluxes through cell boundaries can be directly obtained while conserving physical invariants. These properties result from the geometric approach of linear approximation in tetrahedal cells. Thus the computation of moments of the distribution function such as current densities from Monte-Carlo sampling can be much more efficient and stable than in usual orbit tracers. Furthermore, the mesh is agnostic to whether an orbit is inside or outside the magnetic, and does not rely on flux coordinates. separatrix Therefore `GORILLA` is a flexible tool that is useful in a variety of applications for non-axisymmetric devices, in particular stellarators and tokamaks with 3D perturbations.

In addition we cite now another paper of Kasilov et al (2016) that demonstrates a similar but simpler approach in 2D axisymmetric geometry.
