### 2023-05-31 #4890 @pibion
@krystophny this looks great.  My only remaining comment is on lines 71-72 it looks like there might be some missing/misplaced words?

"inside or outside the magnetic <core>"?

Not sure what "separatrix" is referring to but it looks like a sentence fragment.

Once you address these two lines go ahead and make a release and mint your DOI and I'll recommend for publication!
