+++ PAPER_JOSS/paper.md
@@ -23,7 +23,7 @@ authors:
   - name: Daniel Forstenlechner
     affiliation: 1
   - name: Georg S. Graßler
-    affiliation: 1   
+    affiliation: 1
   - name:  Sergei V. Kasilov
     affiliation: "1, 2, 3"
   - name:  Winfried Kernbichler
@@ -53,27 +53,27 @@ Extremely hot plasmas with a temperature of the order of hundred million degrees
 An important kinetic approach for simulating the collective behavior of a plasma uses direct modeling of particle orbits. A well-known approximation for computing the motion of electrically charged particles in slowly varying electromagnetic fields is to reduce the dynamical equations by separating the relatively fast circular motion around a point called the guiding-center, and primarily treat the relatively slow drift motion of this point. This drift motion is described by the guiding-center equations in non-canonical Hamiltonian form; see, e.g., [@boozer_guiding_1980], [@littlejohn_variational_1983] and [@cary_hamiltonian_2009].
 
 Here, we provide `GORILLA`: an efficient code for the purpose of solving the guiding-center equations. This code is a numerical implementation of the novel, quasi-geometric integration method in 3D geometry described by @eder_quasi-geometric_2020. A related method for axisymmetric geometry has been realized in the code K2D [@kasilov_geometric_2016].
-Apart from this quasi-geometric approach there exists a variety of codes for tracing guiding-center drift motion including ORBIT [@white_hamiltonian_1984], DCOM [@wakasa2001monte], MOCA [@tribaldos_monte_2001], FORTEC-3D [@satake_non-local_2005], VENUS [@isaev_venusf_2006], NEO-MC [@allmaier_variance_2008], XGC [@ku_full-f_2009], ANTS [@drevlak_thermal_2009], ASCOT [@kurki-suonio_ascot_2009], VENUS-LEVIS [@pfefferle_venus-levis_2014], SIMPLE [@albert_symplectic_2020] etc; see also [@beidler_benchmarking_2011]. Most of these codes solve the guiding-center equations primarily in the plasma core of toroidal fusion devices. Due to its formulation in general curvilinear coordinates, `GORILLA` is not limited by the field topology. That means that the computation domain of `GORILLA` covers both the closed field line region (i.e. the plasma core) and the open field line region (i.e. the scrape-off layer). 
+Apart from this quasi-geometric approach there exists a variety of codes for tracing guiding-center drift motion including ORBIT [@white_hamiltonian_1984], DCOM [@wakasa2001monte], MOCA [@tribaldos_monte_2001], FORTEC-3D [@satake_non-local_2005], VENUS [@isaev_venusf_2006], NEO-MC [@allmaier_variance_2008], XGC [@ku_full-f_2009], ANTS [@drevlak_thermal_2009], ASCOT [@kurki-suonio_ascot_2009], VENUS-LEVIS [@pfefferle_venus-levis_2014], SIMPLE [@albert_symplectic_2020] etc; see also [@beidler_benchmarking_2011]. Most of these codes solve the guiding-center equations primarily in the plasma core of toroidal fusion devices. Due to its formulation in general curvilinear coordinates, `GORILLA` is not limited by the field topology. That means that the computation domain of `GORILLA` covers both the closed field line region (i.e. the plasma core) and the open field line region (i.e. the scrape-off layer).
 
 # Summary
 
-`GORILLA` is a Fortran code that computes guiding-center orbits for charged particles of given mass, charge and energy in toroidal fusion devices with three-dimensional field geometry. 
+`GORILLA` is a Fortran code that computes guiding-center orbits for charged particles of given mass, charge and energy in toroidal fusion devices with three-dimensional field geometry.
[…]
