### 2023-06-08 #4890 @krystophny
@pibion thanks, all done. The DOI of the latest release is now 10.5281/zenodo.8017046 and the "all versions" DOI 10.5281/zenodo.4593661 . Let me know if you need anything else.
