### 2023-06-19 #4890 @pibion
@krystophny argh, apologies, an invalid DOI snuck past me.  Could you fix this and release a v1.1.2?
