### 2023-06-19 #4890 @pibion
I've taken a look at the new pdf and it looks great, no additional work is needed there.
