+++ PAPER_JOSS/paper.bib
@@ -324,7 +324,7 @@ volume = {207},
 pages = {282-286},
 year = {2016},
 issn = {0010-4655},
-doi = {https://doi.org/10.1016/j.cpc.2016.07.019},
+doi = {10.1016/j.cpc.2016.07.019},
 url = {https://www.sciencedirect.com/science/article/pii/S0010465516302077},
 author = {S.V. Kasilov and A.M. Runov and W. Kernbichler},
 keywords = {Plasma physics, Kinetic modelling, Numerical integrators},
