### 2023-06-20 #4890 @krystophny
@pibion thanks and sorry. Latest release works with zenodo DOI 10.5281/zenodo.8059175
