### 2023-06-27 9005723
Typographic edits to JOSS paper
